@@ -1,64 +1,61 @@
 {
-    "name": "portal",
-    "author": "Ioannis Tsiakkas",
-    "version": "0.7.0",
-    "private": true,
-    "main": "src/index.ts",
-    "scripts": {
-        "build": "npx tsc",
-        "start": "node build/index.js",
-        "build-start": "npx tsc && node build/index.js",
-        "test": "npx tsc && jest",
-        "lint": "npx eslint . --ext .ts --fix"
-    },
-    "dependencies": {
-        "@discordjs/builders": "^1.1.0",
-        "@discordjs/opus": "^0.8.0",
-        "@discordjs/rest": "^1.0.1",
-        "@discordjs/voice": "^0.11.0",
-        "@types/jsonfile": "^6.1.0",
-        "cheerio": "^1.0.0-rc.12",
-        "core-util-is": "^1.0.3",
-        "discord-api-types": "^0.37.0",
-        "discord-ytdl-core": "^5.0.4",
-        "discord.js": "^13.6.0",
-        "dotenv": "^16.0.1",
-        "eslint-config-prettier": "^8.5.0",
-        "ffmpeg-static": "^5.0.2",
-        "give-me-a-joke": "^0.5.1",
-        "jest": "^28.1.3",
-        "jsonfile": "^6.1.0",
-        "lodash": "^4.17.21",
-        "moment": "^2.29.3",
-        "mongoose": "^6.5.0",
-        "openai": "^3.0.0",
-        "optionator": "^0.9.1",
-        "opusscript": "0.0.8",
-        "roll": "^1.3.1",
-        "translate": "^1.4.1",
-        "voca": "^1.4.0",
-        "winston": "^3.8.1",
-        "yt-search": "^2.10.3",
-        "ytdl-core": "^4.11.0"
-    },
-    "devDependencies": {
-        "@types/jest": "^28.1.6",
-        "@types/lodash": "^4.14.182",
-        "@types/node": "^18.6.3",
-        "@types/node-fetch": "2.6.2",
-        "@types/roll": "^1.2.0",
-        "@types/voca": "^1.4.2",
-        "@types/ws": "^8.5.3",
-        "@types/yt-search": "^2.3.2",
-        "@typescript-eslint/eslint-plugin": "^5.31.0",
-        "@typescript-eslint/parser": "^5.31.0",
-        "eslint": "^8.20.0",
-        "eslint-plugin-prettier": "^4.2.1",
-        "nodemon": "^2.0.15",
-        "ts-jest": "^28.0.7",
-        "typescript": "^4.7.4"
-    },
-    "optionalDependencies": {
-        "fsevents": "^2.3.2"
-    }
+  "name": "portal",
+  "author": "Ioannis Tsiakkas",
+  "version": "0.7.0",
+  "private": true,
+  "main": "src/index.ts",
+  "scripts": {
+    "build": "npx tsc",
+    "start": "node build/index.js",
+    "build-start": "npx tsc && node build/index.js",
+    "test": "npx tsc && jest",
+    "lint": "eslint --ext .ts .",
+    "lint-fix": "npm run lint -- --fix",
+    "dev": "tsc --watch"
+  },
+  "dependencies": {
+    "@discordjs/opus": "^0.9.0",
+    "@discordjs/voice": "^0.16.0",
+    "@types/jsonfile": "^6.1.1",
+    "cheerio": "^1.0.0-rc.12",
+    "core-util-is": "^1.0.3",
+    "dayjs": "^1.11.7",
+    "discord-ytdl-core": "^5.0.4",
+    "discord.js": "^14.11.0",
+    "dotenv": "^16.0.3",
+    "eslint-config-prettier": "^8.8.0",
+    "ffmpeg-static": "^5.1.0",
+    "jest": "^29.5.0",
+    "jsonfile": "^6.1.0",
+    "lodash": "^4.17.21",
+    "mongoose": "^7.1.0",
+    "openai": "^3.2.1",
+    "optionator": "^0.9.1",
+    "opusscript": "0.1.0",
+    "roll": "^1.3.1",
+    "voca": "^1.4.0",
+    "winston": "^3.8.2",
+    "yt-search": "^2.10.4",
+    "ytdl-core": "^4.11.4"
+  },
+  "devDependencies": {
+    "@types/jest": "^29.5.1",
+    "@types/lodash": "^4.14.194",
+    "@types/node": "^20.1.1",
+    "@types/node-fetch": "2.6.3",
+    "@types/roll": "^1.2.0",
+    "@types/voca": "^1.4.2",
+    "@types/ws": "^8.5.4",
+    "@types/yt-search": "^2.3.2",
+    "@typescript-eslint/eslint-plugin": "^5.59.6",
+    "@typescript-eslint/parser": "^5.59.6",
+    "eslint": "^8.41.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "nodemon": "^2.0.22",
+    "ts-jest": "^29.1.0",
+    "typescript": "^5.0.4"
+  },
+  "optionalDependencies": {
+    "fsevents": "^2.3.2"
+  }
 }
