@@ -1,33 +1,24 @@
-import { Channel, PartialDMChannel, TextChannel, VoiceChannel } from "discord.js";
-import { ChannelTypes } from "discord.js/typings/enums";
-import { PortalChannelTypes } from "../data/enums/PortalChannel.enum";
-import { deleted_channel_sync } from "../libraries/mongo.library";
+import { Channel, ChannelType, PartialDMChannel, TextChannel, VoiceChannel } from 'discord.js';
+import { PortalChannelTypes } from '../types/enums/PortalChannel.enum';
+import { deletedChannelSync } from '../libraries/mongo.library';
 
-module.exports = async (
-	args: { channel: Channel | PartialDMChannel }
-): Promise<string> => {
-	return new Promise((resolve, reject) => {
-		if (args.channel.type !== ChannelTypes.GUILD_TEXT.toString() &&
-			args.channel.type !== ChannelTypes.GUILD_VOICE.toString()) {
-			return reject(`only voice and text channels are handled`);
-		}
+export default async function channelDelete(args: { channel: Channel | PartialDMChannel }): Promise<string> {
+  if (args.channel.type !== ChannelType.GuildText && args.channel.type !== ChannelType.GuildVoice) {
+    throw new Error('only voice and text channels are handled');
+  }
 
-		const current_channel = (typeof args.channel === typeof VoiceChannel)
-			? <VoiceChannel>args.channel
-			: <TextChannel>args.channel;
+  const currentChannel = typeof args.channel === typeof VoiceChannel
+    ? args.channel as VoiceChannel
+    : args.channel as TextChannel;
 
-		deleted_channel_sync(current_channel)
-			.then(r => {
-				if (r > 0) {
-					return resolve(`${PortalChannelTypes[r].toString()} channel removed from ` +
-						`${current_channel.guild.name}|${current_channel.guild.id}`);
-				} else {
-					return resolve(`${current_channel.name} channel is not controlled by Portal`);
-				}
-			})
-			.catch(e => {
-				return reject(`error syncing deleted channel from ` +
-					`${current_channel.guild.name}|${current_channel.guild.id}: ${e}`);
-			});
-	});
-};
+  const deletedPChannel = await deletedChannelSync(currentChannel);
+
+  if (!deletedPChannel) {
+    throw new Error(`error syncing deleted channel from ${currentChannel.guild.name}|${currentChannel.guild.id}`);
+  }
+
+  return deletedPChannel > 0
+    ? `${PortalChannelTypes[deletedPChannel].toString()} channel removed from ` +
+    `${currentChannel.guild.name}|${currentChannel.guild.id}`
+    : `${currentChannel.name} channel is not controlled by Portal`;
+}
