@@ -1,8 +1,7 @@
 module.exports = {
-    semi: true,
-    trailingComma: "es5",
-    singleQuote: true,
-    printWidth: 120,
-    tabWidth: 4,
-    useTabs: false
+  semi: true,
+  singleQuote: true,
+  printWidth: 120,
+  tabWidth: 2,
+  useTabs: false,
 };
