@@ -1,68 +1,63 @@
-import { Message } from "discord.js";
-import { included_in_ignore_list } from "../../libraries/guild.library";
-import { messageHelp } from "../../libraries/help.library";
-import { insert_ignore, remove_ignore } from "../../libraries/mongo.library";
-import { GuildPrtl } from "../../types/classes/GuildPrtl.class";
-import { ReturnPormise } from "../../types/classes/TypesPrtl.interface";
 import { SlashCommandBuilder } from '@discordjs/builders';
+import { ChatInputCommandInteraction } from 'discord.js';
+import { includedInPIgnores } from '../../libraries/guild.library';
+import { insertIgnore, removeIgnore } from '../../libraries/mongo.library';
+import { PGuild } from '../../types/classes/PGuild.class';
+import { ReturnPromise } from '../../types/classes/PTypes.interface';
+import { commandDescriptionByNameAndAuthenticationLevel } from '../../libraries/help.library';
 
-module.exports = {
-    data: new SlashCommandBuilder()
-        .setName('ignore')
-        .setDescription('ignore user or channel from spam'),
-    async execute(
-        message: Message, args: string[], guild_object: GuildPrtl
-    ): Promise<ReturnPormise> {
-        return new Promise((resolve) => {
-            if (!message.guild) {
-                return resolve({
-                    result: false,
-                    value: 'guild could not be fetched'
-                });
-            }
+const COMMAND_NAME = 'ignore';
 
-            if (args.length === 0) { // channel ignore
-                if (included_in_ignore_list(message.channel.id, guild_object)) {
-                    remove_ignore(guild_object.id, message.channel.id)
-                        .then(r => {
-                            return resolve({
-                                result: r,
-                                value: r
-                                    ? 'successfully removed ignore channel'
-                                    : 'failed to remove ignore channel'
-                            });
-                        })
-                        .catch(() => {
-                            return resolve({
-                                result: false,
-                                value: 'failed to remove ignore channel'
-                            });
-                        });
-                }
-                else {
-                    insert_ignore(guild_object.id, message.channel.id)
-                        .then(r => {
-                            return resolve({
-                                result: r,
-                                value: r
-                                    ? 'set as an ignore channel successfully'
-                                    : 'failed to set as an ignore channel'
-                            });
-                        })
-                        .catch(e => {
-                            return resolve({
-                                result: false,
-                                value: `failed to set as an ignore channel: ${e}`
-                            });
-                        });
-                }
-            }
-            else {
-                return resolve({
-                    result: false,
-                    value: messageHelp('commands', 'ignore')
-                });
-            }
-        });
+export = {
+  data: new SlashCommandBuilder()
+    .setName(COMMAND_NAME)
+    .setDescription(commandDescriptionByNameAndAuthenticationLevel(COMMAND_NAME, true)),
+  async execute(interaction: ChatInputCommandInteraction, pGuild: PGuild): Promise<ReturnPromise> {
+    if (!interaction.guild) {
+      return {
+        result: false,
+        value: 'guild could not be fetched',
+      };
     }
+
+    const channelId = interaction.channel?.id;
+
+    if (!channelId) {
+      return {
+        result: false,
+        value: 'could not fetch channel',
+      };
+    }
+
+    // channel ignore
+    if (includedInPIgnores(channelId, pGuild)) {
+      const response = await removeIgnore(pGuild.id, channelId);
+
+      if (response) {
+        return {
+          result: response,
+          value: response ? 'successfully removed ignore channel' : 'failed to remove ignore channel',
+        };
+      } else {
+        return {
+          result: false,
+          value: 'failed to remove ignore channel',
+        };
+      }
+    } else {
+      const response = await insertIgnore(pGuild.id, channelId);
+
+      if (response) {
+        return {
+          result: response,
+          value: response ? 'set as an ignore channel successfully' : 'failed to set as an ignore channel',
+        };
+      } else {
+        return {
+          result: false,
+          value: 'failed to set as an ignore channel',
+        };
+      }
+    }
+  },
 };
