@@ -1,119 +1,118 @@
-import { Message } from "discord.js";
-import { askForApproval, isMod, messageHelp } from "../../libraries/help.library";
-import { kick } from "../../libraries/user.library";
-import { ReturnPormise } from "../../types/classes/TypesPrtl.interface";
 import { SlashCommandBuilder } from '@discordjs/builders';
+import { ChatInputCommandInteraction } from 'discord.js';
+import { ReturnPromise } from '../../types/classes/PTypes.interface';
+import { commandDescriptionByNameAndAuthenticationLevel } from '../../libraries/help.library';
 
-module.exports = {
-    data: new SlashCommandBuilder()
-        .setName('kick')
-        .setDescription('kick a user'),
-    async execute(
-        message: Message, args: string[]
-    ): Promise<ReturnPormise> {
-        return new Promise((resolve) => {
-            if (!message.member) {
-                return resolve({
-                    result: false,
-                    value: 'message author could not be fetched'
-                });
-            }
+const COMMAND_NAME = 'kick';
 
-            if (!isMod(message.member)) {
-                return resolve({
-                    result: false,
-                    value: `you must be a Portal moderator to ban users`
-                });
-            }
+export = {
+  data: new SlashCommandBuilder()
+    .setName(COMMAND_NAME)
+    .setDescription(commandDescriptionByNameAndAuthenticationLevel(COMMAND_NAME, true)),
+  async execute(interaction: ChatInputCommandInteraction): Promise<ReturnPromise> {
+    return {
+      result: false,
+      value: 'not yet implemented',
+    };
 
-            if (!message.guild) {
-                return resolve({
-                    result: false,
-                    value: `user guild could not be fetched`
-                });
-            }
+    // const member = interaction.member as GuildMember;
+    // if (!member) {
+    //   return {
+    //     result: false,
+    //     value: 'message author could not be fetched',
+    //   };
+    // }
 
-            let kick_reason = args
-                .join(' ')
-                .substr(args.join(' ').indexOf('|') + 1)
-                .replace(/\s/g, ' ')
-                .trim();
+    // if (!isMod(member)) {
+    //   return {
+    //     result: false,
+    //     value: `you must be a Portal moderator to ban users`,
+    //   };
+    // }
 
-            if (kick_reason === '') {
-                kick_reason = 'kicked by admin'
-            }
+    // if (!interaction.guild) {
+    //   return {
+    //     result: false,
+    //     value: `user guild could not be fetched`,
+    //   };
+    // }
 
-            if (message.mentions && message.mentions.members) {
-                if (message.mentions.members.size === 0) {
-                    return resolve({
-                        result: false,
-                        value: messageHelp('commands', 'kick', `you must tag a member`)
-                    });
-                }
+    // let kickReason = args
+    //   .join(' ')
+    //   .substring(args.join(' ').indexOf('|') + 1)
+    //   .replace(/\s/g, ' ')
+    //   .trim();
 
-                const member_to_kick = message.mentions.members.first() || message.guild.members.cache.get(args[0]);
+    // if (kickReason === '') {
+    //   kickReason = 'kicked by admin';
+    // }
 
-                if (member_to_kick) {
-                    if (message.member === member_to_kick) {
-                        return resolve({
-                            result: false,
-                            value: messageHelp('commands', 'kick', `you can't kick on yourself`)
-                        });
-                    }
+    // if (interaction.mentions && interaction.mentions.members) {
+    //   if (interaction.mentions.members.size === 0) {
+    //     return {
+    //       result: false,
+    //       value: messageHelp('commands', 'kick', `you must tag a member`),
+    //     };
+    //   }
 
-                    askForApproval(
-                        message,
-                        message.member,
-                        `*${message.member}, are you sure you want to kick ` +
-                        `user ${member_to_kick}*, do you **(yes / no)** ?`
-                    )
-                        .then(result => {
-                            if (result) {
-                                kick(member_to_kick, kick_reason)
-                                    .then(r => {
-                                        return resolve({
-                                            result: r,
-                                            value: r
-                                                ? `${member_to_kick} has been kicked by ${message.author} ` +
-                                                `because: *${kick_reason}*`
-                                                : `${member_to_kick} is not kickable`
-                                        });
-                                    })
-                                    .catch(e => {
-                                        return resolve({
-                                            result: false,
-                                            value: `failed to kick member ${member_to_kick}, ` +
-                                                `Portal's role must be higher than member you want to kick: ${e}`
-                                        });
-                                    });
-                            } else {
-                                return resolve({
-                                    result: false,
-                                    value: `user ${member_to_kick} will not be kicked`
-                                });
-                            }
-                        })
-                        .catch(e => {
-                            return resolve({
-                                result: false,
-                                value: `failed to kick: ${e}`
-                            });
-                        });
-                }
-                else {
-                    return resolve({
-                        result: false,
-                        value: `could not find member`
-                    });
-                }
+    //   const memberToKick = interaction.mentions.members.first() || interaction.guild.members.cache.get(args[0]);
 
-            } else {
-                return resolve({
-                    result: false,
-                    value: `no user mentioned to kick`
-                });
-            }
-        });
-    }
+    //   if (memberToKick) {
+    //     if (member === memberToKick) {
+    //       return {
+    //         result: false,
+    //         value: messageHelp('commands', 'kick', `you can't kick on yourself`),
+    //       };
+    //     }
+
+    //     askForApproval(
+    //       interaction,
+    //       member,
+    //       `*${member}, are you sure you want to kick ` + `user ${memberToKick}*, do you **(yes / no)** ?`
+    //     )
+    //       .then((result) => {
+    //         if (result) {
+    //           kick(memberToKick, kickReason)
+    //             .then((r) => {
+    //               return {
+    //                 result: r,
+    //                 value: r
+    //                   ? `${memberToKick} has been kicked by ${interaction.author} ` + `because: *${kickReason}*`
+    //                   : `${memberToKick} is not kickable`,
+    //               };
+    //             })
+    //             .catch((e) => {
+    //               return {
+    //                 result: false,
+    //                 value:
+    //                   `failed to kick member ${memberToKick}, ` +
+    //                   `Portal's role must be higher than member you want to kick: ${e}`,
+    //               };
+    //             });
+    //         } else {
+    //           return {
+    //             result: false,
+    //             value: `user ${memberToKick} will not be kicked`,
+    //           };
+    //         }
+    //       })
+    //       .catch((e) => {
+    //         return {
+    //           result: false,
+    //           value: `failed to kick: ${e}`,
+    //         };
+    //       });
+    //   } else {
+    //     return {
+    //       result: false,
+    //       value: `could not find member`,
+    //     };
+    //   }
+    // } else {
+    //   return {
+    //     result: false,
+    //     value: `no user mentioned to kick`,
+    //   };
+    // }
+  },
 };
-
