@@ -1,102 +1,105 @@
-import { GuildChannelCreateOptions, Message } from "discord.js";
-import { ChannelTypes } from "discord.js/typings/enums";
-import { create_channel } from "../../libraries/guild.library";
-import { messageHelp } from "../../libraries/help.library";
-import { insert_portal } from "../../libraries/mongo.library";
-import { GuildPrtl } from "../../types/classes/GuildPrtl.class";
-import { IPortalChannelPrtl, PortalChannelPrtl } from "../../types/classes/PortalChannelPrtl.class";
-import { ReturnPormise } from "../../types/classes/TypesPrtl.interface";
 import { SlashCommandBuilder } from '@discordjs/builders';
+import { ChatInputCommandInteraction, GuildMember, VoiceChannel } from 'discord.js';
+import { commandDescriptionByNameAndAuthenticationLevel, messageHelp } from '../../libraries/help.library';
+import { insertPortal } from '../../libraries/mongo.library';
+import { PGuild } from '../../types/classes/PGuild.class';
+import { IPChannel, PChannel } from '../../types/classes/PPortalChannel.class';
+import { ReturnPromise } from '../../types/classes/PTypes.interface';
 
-module.exports = {
-    data: new SlashCommandBuilder()
-        .setName('portal')
-        .setDescription('create portal channel'),
-    async execute(
-        message: Message, args: string[], guild_object: GuildPrtl
-    ): Promise<ReturnPormise> {
-        return new Promise((resolve) => {
-            if (args.length === 0) {
-                return resolve({
-                    result: false,
-                    value: messageHelp('commands', 'portal')
-                });
-            }
+const COMMAND_NAME = 'portal';
 
-            if (!message.guild) {
-                return resolve({
-                    result: true,
-                    value: 'guild could not be fetched'
-                });
-            }
+export = {
+  data: new SlashCommandBuilder()
+    .setName(COMMAND_NAME)
+    .setDescription(commandDescriptionByNameAndAuthenticationLevel(COMMAND_NAME, true))
+    .addChannelOption((option) =>
+      option
+        .setName('portal_channel_name')
+        .setDescription('the name of the portal channel you want to create')
+        .setRequired(true))
+    .setDMPermission(false),
+  async execute(interaction: ChatInputCommandInteraction, pGuild: PGuild): Promise<ReturnPromise> {
+    const newPortalChannel = interaction.options.getChannel('portal_channel_name');
 
-            if (!message.member) {
-                return resolve({
-                    result: true,
-                    value: 'member could not be fetched'
-                });
-            }
-
-            const current_guild = message.guild;
-            const current_member = message.member;
-
-            let portal_channel: string = args.join(' ').substr(0, args.join(' ').indexOf('|'));
-            let portal_category: string | null = args.join(' ').substr(args.join(' ').indexOf('|') + 1);
-
-            if (portal_channel === '' && portal_category !== '') {
-                portal_channel = portal_category;
-                portal_category = null;
-            }
-
-            const portal_options: GuildChannelCreateOptions = {
-                topic: `by Portal, channels on demand`,
-                type: ChannelTypes.GUILD_VOICE,
-                bitrate: 32000,
-                userLimit: 1
-            };
-
-            const voice_regex = guild_object.premium
-                // ? 'G$#-P$member_count | $status_list'
-                ? `$#:$member_count {{
-                "if": "$status_count", "is": "===", "with": "1",
-                "yes": "$status_list", "no": "$status_list|acronym"
-            }}`
-                : 'Channel $#';
-
-            create_channel(current_guild, portal_channel, portal_options, portal_category)
-                .then(r_channel => {
-                    const new_portal = new PortalChannelPrtl(r_channel, current_member.id,
-                        true, portal_channel, voice_regex, [], false, null, guild_object.locale, true, true, 0, false);
-
-                    insert_portal(guild_object.id, new_portal as IPortalChannelPrtl)
-                        .then(r_portal => {
-                            if (r_portal) {
-                                return resolve({
-                                    result: true,
-                                    value: 'portal channel has been created.\n' +
-                                        'Keep in mind that due to Discord\'s limitations,\n' +
-                                        'channel names will be updated on a five minute interval'
-                                });
-                            } else {
-                                return resolve({
-                                    result: false,
-                                    value: 'portal channel failed to be created'
-                                });
-                            }
-                        })
-                        .catch(e => {
-                            return resolve({
-                                result: false,
-                                value: `portal channel failed to be created: ${e}`
-                            });
-                        });
-                })
-                .catch(e => {
-                    return resolve({
-                        result: false,
-                        value: `an error occurred while creating channel: ${e}`
-                    });
-                });
-        });
+    if (!newPortalChannel) {
+      return {
+        result: false,
+        value: messageHelp('commands', 'portal', 'portal channel name is required'),
+      };
     }
+
+    if (!(newPortalChannel instanceof VoiceChannel)) {
+      return {
+        result: false,
+        value: messageHelp('commands', 'portal', 'channel must be voice channel'),
+      };
+    }
+
+    if (!interaction.guild) {
+      return {
+        result: true,
+        value: 'guild could not be fetched',
+      };
+    }
+
+    if (!interaction.member) {
+      return {
+        result: true,
+        value: 'member could not be fetched',
+      };
+    }
+
+    // const currentGuild = interaction.guild as Guild;
+    const currentMember = interaction.member as GuildMember;
+
+    // const portalOptions: GuildChannelCreateOptions = {
+    //   name: 'portal',
+    //   topic: `by Portal, channels on demand`,
+    //   type: ChannelType.GuildVoice,
+    //   bitrate: 32000,
+    //   userLimit: 1,
+    // };
+
+    const voiceRegex = pGuild.premium
+      ? // ? 'G$#-P$memberCount | $statusList'
+      '$#:$memberCount {{"if": "$statusCount", "is": "===", "with": "1","yes": "$statusList", "no": "$statusList|acronym"}}'
+      : 'Channel $#';
+
+    // const newPortalChannelId = await createChannel(currentGuild, portalChannelName, portalOptions, portalChannelCategoryName);
+
+    // if (!newPortalChannelId) {
+    //   return {
+    //     result: false,
+    //     value: `an error occurred while creating channel}`,
+    //   };
+    // }
+
+    const pChannel = new PChannel(
+      newPortalChannel.id,
+      currentMember.id,
+      true,
+      'portal',
+      voiceRegex,
+      [],
+      false,
+      null,
+      pGuild.locale,
+      true,
+      true,
+      0,
+      false
+    );
+
+    const portalInserted = await insertPortal(pGuild.id, pChannel as unknown as IPChannel);
+
+    return {
+      result: !!portalInserted,
+      value:
+        portalInserted ? 'Portal channel has been created.\n' +
+          'Keep in mind that due to Discord\'s limitations,\n' +
+          'channel names will be updated on a five minute interval' :
+          'portal channel failed to be created',
+    };
+
+  },
 };
