@@ -1,56 +1,58 @@
-import { Message, VoiceChannel } from "discord.js";
-import { messageHelp } from "../../libraries/help.library";
-import { GuildPrtl } from "../../types/classes/GuildPrtl.class";
-import { ReturnPormise } from "../../types/classes/TypesPrtl.interface";
-import { set_attribute } from "../../types/interfaces/Attribute.interface";
 import { SlashCommandBuilder } from '@discordjs/builders';
+import { ChatInputCommandInteraction } from 'discord.js';
+import { PGuild } from '../../types/classes/PGuild.class';
+import { ReturnPromise } from '../../types/classes/PTypes.interface';
+import { commandDescriptionByNameAndAuthenticationLevel } from '../../libraries/help.library';
 
-module.exports = {
-    data: new SlashCommandBuilder()
-        .setName('set')
-        .setDescription('set the value of an attribute'),
-    async execute(
-        message: Message, args: string[], guild_object: GuildPrtl
-    ): Promise<ReturnPormise> {
-        return new Promise((resolve) => {
-            if (!message.guild) {
-                return resolve({
-                    result: true,
-                    value: 'guild could not be fetched'
-                });
-            }
+const COMMAND_NAME = 'set';
 
-            if (!message.member) {
-                return resolve({
-                    result: true,
-                    value: 'member could not be fetched'
-                });
-            }
+export = {
+  data: new SlashCommandBuilder()
+    .setName(COMMAND_NAME)
+    .setDescription(commandDescriptionByNameAndAuthenticationLevel(COMMAND_NAME, true)),
+  async execute(interaction: ChatInputCommandInteraction, pGuild: PGuild): Promise<ReturnPromise> {
+    return {
+      result: true,
+      value: 'not yet implemented',
+    };
 
-            if (args.length >= 2) {
-                const value_array = [...args];
-                value_array.shift();
+    // if (!interaction.guild) {
+    //   return {
+    //     result: true,
+    //     value: 'guild could not be fetched',
+    //   };
+    // }
 
-                const value = value_array
-                    .filter(val => val !== '\n')
-                    .join(' ');
+    // const member = interaction.member as GuildMember;
+    // if (!member) {
+    //   return {
+    //     result: true,
+    //     value: 'member could not be fetched',
+    //   };
+    // }
 
-                set_attribute(message.member.voice.channel as VoiceChannel, guild_object, args[0], value, message.member, message)
-                    .then(r => {
-                        return resolve(r);
-                    })
-                    .catch(e => {
-                        return resolve({
-                            result: false,
-                            value: `something went wrong in set function: ${e}`
-                        });
-                    });
-            } else {
-                return resolve({
-                    result: false,
-                    value: messageHelp('commands', 'set', 'arguments are set by name and value')
-                });
-            }
-        });
-    }
+    // if (args.length >= 2) {
+    //   const valueArray = [...args];
+    //   valueArray.shift();
+
+    //   const value = valueArray.filter((val) => val !== '\n').join(' ');
+
+    //   // !this must be updated to handle interaction instead of message, but needs restructure not only refactoring
+    //   // setAttribute(member.voice.channel as VoiceChannel, pGuild, args[0], value, member, interaction)
+    //   //   .then((r) => {
+    //   //     return r;
+    //   //   })
+    //   //   .catch((e) => {
+    //   //     return {
+    //   //       result: false,
+    //   //       value: `something went wrong in set function: ${e}`,
+    //   //     };
+    //   //   });
+    // } else {
+    //   return {
+    //     result: false,
+    //     value: messageHelp('commands', 'set', 'arguments are set by name and value'),
+    //   };
+    // }
+  },
 };
