@@ -1,440 +1,407 @@
-import { getVoiceConnection, VoiceConnection } from "@discordjs/voice";
-import { Client, Guild, TextChannel, VoiceChannel, VoiceState } from "discord.js";
-import { create_voice_channel, generate_channel_name, included_in_portal_list, included_in_voice_list } from "../libraries/guild.library";
-import { isChannelDeleted, isGuildDeleted, logger, updateMusicLyricsMessage, updateMusicMessage } from "../libraries/help.library";
-// import { client_talk } from "../libraries/localisation.library";
-import { fetch_guild, remove_voice, set_music_data, updateGuild } from "../libraries/mongo.library";
-import { update_timestamp } from "../libraries/user.library";
-import { GuildPrtl } from "../types/classes/GuildPrtl.class";
-import { PortalChannelPrtl } from "../types/classes/PortalChannelPrtl.class";
+import { getVoiceConnection } from '@discordjs/voice';
+import { Client, Guild, TextChannel, VoiceChannel, VoiceState } from 'discord.js';
+import {
+  createVoiceChannel,
+  generateChannelName,
+  includedInPChannels,
+  includedInVoiceList,
+} from '../libraries/guild.library';
+import {
+  isChannelDeleted,
+  isGuildDeleted,
+  logger,
+  updateMusicLyricsMessage,
+  updateMusicMessage,
+} from '../libraries/help.library';
+import { fetchGuild, removeVoice, setMusicData, updateGuild } from '../libraries/mongo.library';
+import { updateTimestamp } from '../libraries/user.library';
+import { PGuild } from '../types/classes/PGuild.class';
+import { PChannel } from '../types/classes/PPortalChannel.class';
+
+export default async function (args: { client: Client; newState: VoiceState; oldState: VoiceState }): Promise<string> {
+  if (!args.newState?.guild) {
+    return 'could not find guild in Portal';
+  }
+
+  const newChannel = args.newState.channel; // join channel
+  const oldChannel = args.oldState.channel; // left channel
+
+  const pGuild = await fetchGuild(args.newState?.guild.id);
+
+  if (!pGuild) {
+    return 'could not find guild in Portal';
+  }
+
+  if (!pGuild) {
+    return 'could not find guild in Portal';
+  }
+
+  if (newChannel) {
+    for (let i = 0; i < pGuild.pChannels.length; i++) {
+      const p = pGuild.pChannels[i];
+
+      if (p.id === newChannel.id) {
+        if (p.noBots && args.newState.member?.user.bot) {
+          args.newState.disconnect('voice channel does not allow bots')
+            .catch((e) => {
+              return `failed to kick: ${e}`;
+            });
+
+          channelEmptyCheck(newChannel as VoiceChannel, pGuild, args.client)
+            .catch((e) => {
+              logger.error(new Error(`failed to check channel state: ${e}`));
+            });
+
+          return 'portal channel does not allow bots';
+        }
+      }
+
+      for (let i = 0; i < p.pVoiceChannels.length; i++) {
+        const v = p.pVoiceChannels[i];
+
+        if (v.id === newChannel.id) {
+          if (v.noBots) {
+            args.newState.disconnect('voice channel does not allow bots')
+              .catch((e) => {
+                return `failed to kick: ${e}`;
+              });
+
+            channelEmptyCheck(newChannel as VoiceChannel, pGuild, args.client)
+              .catch((e) => {
+                logger.error(new Error(`failed to check channel state: ${e}`));
+              });
+
+            return 'voice channel does not allow bots';
+          }
+        }
+      }
+    }
+  }
+
+  // if (args.client.voice && args.newState.member) {
+  //     const newVoiceConnection = args.client.voice.connections
+  //         .find((connection: VoiceConnection) =>
+  //             !!newChannel && connection.channel.id === newChannel.id);
+
+  //     if (newVoiceConnection && !args.newState.member.user.bot) {
+  //         clientTalk(args.client, pGuild, 'userConnected');
+  //     }
+
+  //     const oldVoiceConnection = args.client.voice.connections
+  //         .find((connection: VoiceConnection) =>
+  //             !!oldChannel && connection.channel.id === oldChannel.id);
+
+  //     if (oldVoiceConnection && !args.newState.member.user.bot) {
+  //         clientTalk(args.client, pGuild, 'userDisconnected');
+  //     }
+  // }
+
+  if (!oldChannel) {
+    const response = await fromNull(newChannel as VoiceChannel | null, pGuild, args.newState);
+
+    if (!response) {
+      return 'failed to create voice channel';
+    }
+
+    return response;
+  } else {
+    const response = await fromExisting(oldChannel as VoiceChannel, newChannel as VoiceChannel | null, args.client, pGuild, args.newState);
+
+    if (!response) {
+      return 'failed to create voice channel';
+    }
+
+    return response;
+  }
+}
 
 // delete portal's voice channel
-async function delete_voice_channel(
-    channel: VoiceChannel | TextChannel, guild_object: GuildPrtl
-): Promise<string> {
-    return new Promise((resolve, reject) => {
-        if (!channel.deletable) {
-            return reject(`channel ${channel.name} (${channel.id}) is not deletable`);
-        } else {
-            guild_object.portal_list.some(p =>
-                p.voice_list.some(v => {
-                    if (v.id === channel.id) {
-                        channel
-                            .delete()
-                            .then(deleted_channel => {
-                                remove_voice(guild_object.id, p.id, v.id)
-                                    .then(r => {
-                                        if (r) {
-                                            return resolve(`channel (${deleted_channel.id}) deleted`);
-                                        } else {
-                                            return reject(`channel (${deleted_channel.id}) failed to be deleted`);
-                                        }
-                                    })
-                                    .catch(e => {
-                                        return reject(`channel (${deleted_channel.id}) failed to be delete: ${e}`);
-                                    });
-                            })
-                            .catch(e => {
-                                return reject(`channel ${channel.name} (${channel.id}) failed to be delete: ${e}`);
-                            });
+async function deleteVoiceChannel(channel: VoiceChannel | TextChannel, pGuild: PGuild): Promise<string> {
+  if (!channel.deletable) {
+    return `channel ${channel.name} (${channel.id}) is not deletable`;
+  }
 
-                        return true;
-                    }
+  for (const pChannel of pGuild.pChannels) {
+    for (const pVoiceChannel of pChannel.pVoiceChannels) {
+      if (pVoiceChannel.id === channel.id) {
+        try {
+          const deletedChannel = await channel.delete();
+          const response = await removeVoice(pGuild.id, pChannel.id, pVoiceChannel.id);
 
-                    return false;
-                })
-            );
+          return response ? `channel (${deletedChannel.id}) deleted` : `channel (${deletedChannel.id}) failed to be deleted`;
+        } catch (e) {
+          return `channel ${channel.name} (${channel.id}) failed to be delete`;
         }
-    });
+      }
+    }
+  }
+
+  return `no portal channels found for ${channel.name} (${channel.id})`;
 }
 
-function five_min_refresher(
-    voice_channel: VoiceChannel, portal_list: PortalChannelPrtl[],
-    guild: Guild, minutes: number
-): void {
-    fetch_guild(guild.id)
-        .then(guild_object => {
-            if (guild_object) {
-                generate_channel_name(voice_channel, portal_list, guild_object, guild)
-                    .catch((e: any) => {
-                        logger.error(new Error(`failed to generate channel name: ${e}`));
-                    });
+async function fiveMinuteRefresher(voiceChannel: VoiceChannel, portalList: PChannel[], guild: Guild, minutes: number): Promise<void> {
+  const pGuild = await fetchGuild(guild.id)
+    .catch(() => {
+      logger.error(new Error('failed to fetch guild'));
+    });
 
-                setTimeout(() => {
-                    if (!isGuildDeleted(guild) && !isChannelDeleted(voice_channel)) {
-                        generate_channel_name(voice_channel, portal_list, guild_object, guild)
-                            .catch((e: any) => {
-                                logger.error(new Error(`failed to generate channel name: ${e}`));
-                            });
+  if (!pGuild) {
+    return;
+  }
 
-                        five_min_refresher(voice_channel, portal_list, guild, minutes);
-                    }
-                }, minutes * 60 * 1000);
-            }
-        })
-        .catch((e: any) => {
-            logger.error(new Error(`failed to fetch guild: ${e}`));
+  generateChannelName(voiceChannel, portalList, pGuild, guild)
+    .catch(() => {
+      logger.error(new Error('failed to generate channel name'));
+    });
+
+  setTimeout(() => {
+    if (!isGuildDeleted(guild) && !isChannelDeleted(voiceChannel)) {
+      generateChannelName(voiceChannel, portalList, pGuild, guild)
+        .catch(() => {
+          logger.error(new Error('failed to generate channel name'));
         });
+
+      fiveMinuteRefresher(voiceChannel, portalList, guild, minutes);
+    }
+  }, minutes * 60 * 1000);
 }
 
-async function channel_empty_check(
-    old_channel: VoiceChannel | TextChannel, guild_object: GuildPrtl, client: Client
+async function channelEmptyCheck(
+  oldChannel: VoiceChannel | TextChannel,
+  pGuild: PGuild,
+  client: Client
 ): Promise<string> {
-    return new Promise((resolve, reject) => {
-        if (old_channel.members.size === 0) {
-            if (included_in_voice_list(old_channel.id, guild_object.portal_list)) {
-                delete_voice_channel(old_channel, guild_object)
-                    .then(response => {
-                        return resolve(response);
-                    })
-                    .catch(e => {
-                        return reject(`an error occurred while deleting voice | ${e}`);
-                    });
-            } else {
-                return resolve(`channel is not handled by Portal`);
-            }
-        }
-        else if (old_channel.members.size === 1) {
-            if (!client.voice) {
-                return resolve(`Portal is not connected`);
-            }
+  if (oldChannel.members.size === 0 && includedInVoiceList(oldChannel.id, pGuild.pChannels)) {
+    return await deleteVoiceChannel(oldChannel, pGuild)
+      .catch((e) => {
+        return `an error occurred while deleting voice | ${e}`;
+      });
+  } else if (oldChannel.members.size === 1) {
+    if (!client.voice) {
+      return 'Portal is not connected';
+    }
 
-            const voiceConnection = getVoiceConnection(old_channel.guild.id);
+    const voiceConnection = getVoiceConnection(oldChannel.guild.id);
 
-            if (!voiceConnection) {
-                return resolve(`Portal is not connected`);
-            }
+    if (!voiceConnection) {
+      return 'Portal is not connected';
+    }
 
-            guild_object.music_queue = [];
-            updateGuild(guild_object.id, 'music_queue', guild_object.music_queue)
-                .catch(e => {
-                    return reject(`failed to update guild: ${e}`);
-                });
-            voiceConnection.disconnect();
+    pGuild.musicQueue = [];
+    updateGuild(pGuild.id, 'musicQueue', pGuild.musicQueue)
+      .catch((e) => {
+        return `failed to update guild: ${e}`;
+      });
+    voiceConnection.disconnect();
 
-            if (guild_object.music_data.pinned) {
-                guild_object.music_data.pinned = false;
-                set_music_data(guild_object.id, guild_object.music_data)
-                    .catch(e => {
-                        return reject(`failed to set music data: ${e}`);
-                    });
-            }
+    if (pGuild.musicData.pinned) {
+      pGuild.musicData.pinned = false;
+      setMusicData(pGuild.id, pGuild.musicData)
+        .catch((e) => {
+          return `failed to set music data: ${e}`;
+        });
+    }
 
-            updateMusicMessage(
-                old_channel.guild,
-                guild_object,
-                guild_object.music_queue.length > 0
-                    ? guild_object.music_queue[0]
-                    : undefined,
-                'left last',
-                false
-            )
-                .catch(e => {
-                    return reject(`failed to update music message: ${e}`);
-                });
+    updateMusicMessage(
+      oldChannel.guild,
+      pGuild,
+      pGuild.musicQueue.length > 0 ? pGuild.musicQueue[0] : undefined,
+      'left last',
+      false
+    )
+      .catch((e) => {
+        return `failed to update music message: ${e}`;
+      });
 
-            updateMusicLyricsMessage(old_channel.guild, guild_object, '')
-                .catch(e => {
-                    return reject(`failed to update music lyrics: ${e}`);
-                });
+    updateMusicLyricsMessage(oldChannel.guild, pGuild, '')
+      .catch((e) => {
+        return `failed to update music lyrics: ${e}`;
+      });
 
-            if (included_in_voice_list(old_channel.id, guild_object.portal_list)) {
-                delete_voice_channel(old_channel, guild_object)
-                    .then(response => {
-                        return resolve(response);
-                    })
-                    .catch(e => {
-                        return reject(`an error occurred while deleting voice | ${e}`)
-                    });
-            } else {
-                return resolve('Portal left voice channel');
-            }
+    if (includedInVoiceList(oldChannel.id, pGuild.pChannels)) {
+      return await deleteVoiceChannel(oldChannel, pGuild)
+        .catch((e) => {
+          return `an error occurred while deleting voice | ${e}`;
+        });
+    } else {
+      return 'Portal left voice channel';
+    }
+  }
 
-        }
-    });
+  return 'channel is not handled by Portal';
 }
 
-async function from_null(
-    new_channel: VoiceChannel | null, guild_object: GuildPrtl, newState: VoiceState
+async function fromNull(newChannel: VoiceChannel | null, pGuild: PGuild, newState: VoiceState): Promise<string> {
+  if (!newChannel) {
+    return 'strange, from null to null';
+  }
+
+  // joined from null
+  if (includedInPChannels(newChannel.id, pGuild.pChannels)) {
+    // joined portal channel
+    const pChannel = pGuild.pChannels.find((p) => p.id === newChannel.id);
+
+    if (!pChannel) {
+      return 'null->existing (source: null | dest: portalList) / could not find pChannel';
+    }
+
+    await createVoiceChannel(newState, pChannel)
+      .catch((e) => {
+        return `null->existing (source: null | dest: portalList): ${e}`;
+      });
+
+    const level = await updateTimestamp(newState, pGuild) // points for voice
+      .catch((e) => {
+        logger.error(new Error(`failed to send message: ${e}`));
+      });
+
+    if (level) {
+      newState.member
+        ?.send(`you reached level ${level} in ${newState.guild}!`)
+        .catch((e) => {
+          logger.error(new Error(`failed to send message: ${e}`));
+        });
+    }
+
+    return 'null->existing (source: null | dest: portalList)';
+  } else if (includedInVoiceList(newChannel.id, pGuild.pChannels)) {
+    // joined voice channel
+    fiveMinuteRefresher(newChannel, pGuild.pChannels, newState.guild, 5);
+    const level = await updateTimestamp(newState, pGuild) // points for voice
+      .catch((e) => {
+        logger.error(new Error(`failed to send message: ${e}`));
+      });
+
+    if (level) {
+      newState.member
+        ?.send(`you reached level ${level} in ${newState.guild}!`)
+        .catch((e) => {
+          logger.error(new Error(`failed to send message: ${e}`));
+        });
+    }
+
+    return 'null->existing (source: null | dest: voiceList)';
+  } else {
+    // joined other channel
+    const level = await updateTimestamp(newState, pGuild) // points for voice
+      .catch((e) => {
+        logger.error(new Error(`failed to send message: ${e}`));
+      });
+
+    if (level) {
+      newState.member
+        ?.send(`you reached level ${level} in ${newState.guild}!`)
+        .catch((e) => {
+          logger.error(new Error(`failed to send message: ${e}`));
+        });
+    }
+
+    return 'null->existing (source: null | dest: other channel)';
+  }
+}
+
+async function fromExisting(
+  oldChannel: VoiceChannel,
+  newChannel: VoiceChannel | null,
+  client: Client,
+  pGuild: PGuild,
+  newState: VoiceState
 ): Promise<string> {
-    return new Promise((resolve, reject) => {
-        if (new_channel) { // joined from null
-            if (included_in_portal_list(new_channel.id, guild_object.portal_list)) { // joined portal channel
-                const portal_object = guild_object.portal_list
-                    .find(p => p.id === new_channel.id);
+  if (newChannel === null) {
+    const level = await updateTimestamp(newState, pGuild) // points for voice
+      .catch((e) => {
+        logger.error(new Error(`failed to send message: ${e}`));
+      });
 
-                if (!portal_object) {
-                    return reject('null->existing (source: null | dest: portal_list) / could not find portal_object');
-                }
+    if (level) {
+      newState.member
+        ?.send(`you reached level ${level} in ${newState.guild}!`)
+        .catch((e) => {
+          logger.error(new Error(`failed to send message: ${e}`));
+        });
+    }
 
-                create_voice_channel(newState, portal_object)
-                    .then(() => {
-                        update_timestamp(newState, guild_object) // points for voice
-                            .then(level => {
-                                if (level) {
-                                    newState.member?.send(`you reached level ${level} in ${newState.guild}!`)
-                                        .catch((e: any) => {
-                                            logger.error(new Error(`failed to send message: ${e}`));
-                                        });
-                                }
-                            })
-                            .catch((e: any) => {
-                                logger.error(new Error(`failed to send message: ${e}`));
-                            });
+    channelEmptyCheck(oldChannel, pGuild, client)
+      .catch((e) => {
+        logger.error(new Error(`failed to check channel state: ${e}`));
+      });
 
-                        return resolve('null->existing (source: null | dest: portal_list)');
-                    })
-                    .catch(e => {
-                        return reject(`null->existing (source: null | dest: portal_list): ${e}`);
-                    });
-            }
-            else if (included_in_voice_list(new_channel.id, guild_object.portal_list)) { // joined voice channel
-                five_min_refresher(new_channel, guild_object.portal_list, newState.guild, 5);
-                update_timestamp(newState, guild_object) // points for voice
-                    .then(level => {
-                        if (level) {
-                            newState.member?.send(`you reached level ${level} in ${newState.guild}!`)
-                                .catch((e: any) => {
-                                    logger.error(new Error(`failed to send message: ${e}`));
-                                });
-                        }
-                    })
-                    .catch((e: any) => {
-                        logger.error(new Error(`failed to send message: ${e}`));
-                    });
+    return 'existing->null';
+  }
 
-                return resolve('null->existing (source: null | dest: voice_list)');
-            }
-            else { // joined other channel
-                update_timestamp(newState, guild_object) // points for voice
-                    .then(level => {
-                        if (level) {
-                            newState.member?.send(`you reached level ${level} in ${newState.guild}!`)
-                                .catch((e: any) => {
-                                    logger.error(new Error(`failed to send message: ${e}`));
-                                });
-                        }
-                    })
-                    .catch((e: any) => {
-                        logger.error(new Error(`failed to send message: ${e}`));
-                    });
-
-                return resolve('null->existing (source: null | dest: other channel)');
-            }
-        } else {
-            return reject('strange, from null to null');
-        }
+  // Moved from channel to channel
+  const level = await updateTimestamp(newState, pGuild) // points for voice
+    .catch((e) => {
+      logger.error(new Error(`failed to send message: ${e}`));
     });
-}
 
-async function from_existing(
-    old_channel: VoiceChannel, new_channel: VoiceChannel | null, client: Client,
-    guild_object: GuildPrtl, newState: VoiceState
-): Promise<string> {
-    return new Promise((resolve, reject) => {
-        if (new_channel === null) {
-            update_timestamp(newState, guild_object) // points for voice
-                .then(level => {
-                    if (level) {
-                        newState.member?.send(`you reached level ${level} in ${newState.guild}!`)
-                            .catch((e: any) => {
-                                logger.error(new Error(`failed to send message: ${e}`));
-                            });
-                    }
-                })
-                .catch((e: any) => {
-                    logger.error(new Error(`failed to send message: ${e}`));
-                });
+  if (level) {
+    newState.member
+      ?.send(`you reached level ${level} in ${newState.guild}!`)
+      .catch((e) => {
+        logger.error(new Error(`failed to send message: ${e}`));
+      });
+  }
 
-            channel_empty_check(old_channel, guild_object, client)
-                .catch(e => {
-                    logger.error(new Error(`failed to check channel state: ${e}`));
-                });
+  if (includedInPChannels(oldChannel.id, pGuild.pChannels)) {
+    if (!includedInVoiceList(newChannel.id, pGuild.pChannels)) {
+      return 'not handled by portal';
+    }
 
-            return resolve('existing->null');
-        }
-        else if (new_channel !== null) { // Moved from channel to channel
-            update_timestamp(newState, guild_object) // points for voice
-                .then(level => {
-                    if (level) {
-                        newState.member?.send(`you reached level ${level} in ${newState.guild}!`)
-                            .catch((e: any) => {
-                                logger.error(new Error(`failed to send message: ${e}`));
-                            });
-                    }
-                })
-                .catch((e: any) => {
-                    logger.error(new Error(`failed to send message: ${e}`));
-                });
+    // has been handled before
+    await fiveMinuteRefresher(newChannel, pGuild.pChannels, newState.guild, 5);
 
-            if (included_in_portal_list(old_channel.id, guild_object.portal_list)) {
-                if (included_in_voice_list(new_channel.id, guild_object.portal_list)) { // has been handled before
-                    five_min_refresher(new_channel, guild_object.portal_list, newState.guild, 5);
+    return 'existing->existing (source: portalList | dest: voiceList) / has been handled before';
+  } else if (includedInVoiceList(oldChannel.id, pGuild.pChannels)) {
+    await channelEmptyCheck(oldChannel, pGuild, client)
+      .catch((e) => {
+        logger.error(new Error(`failed to check channel state: ${e}`));
+      });
 
-                    return resolve('existing->existing (source: portal_list | dest: voice_list) / has been handled before');
-                } else {
-                    return resolve('not handled by portal');
-                }
-            }
-            else if (included_in_voice_list(old_channel.id, guild_object.portal_list)) {
-                channel_empty_check(old_channel, guild_object, client)
-                    .catch(e => {
-                        logger.error(new Error(`failed to check channel state: ${e}`));
-                    });
+    if (includedInPChannels(newChannel.id, pGuild.pChannels)) {
+      // moved from voice to portal
+      const pChannel = pGuild.pChannels.find((p) => p.id === newChannel.id);
 
-                if (included_in_portal_list(new_channel.id, guild_object.portal_list)) { // moved from voice to portal
-                    const portal_object = guild_object.portal_list
-                        .find(p => p.id === new_channel.id);
+      if (!pChannel) {
+        return 'could not find Portal in database';
+      }
 
-                    if (!portal_object) {
-                        return reject('could not find Portal in database');
-                    }
+      await createVoiceChannel(newState, pChannel)
+        .catch((e) => {
+          return `an error occurred while creating voice channel: ${e}`;
+        });
 
-                    create_voice_channel(newState, portal_object)
-                        .then(() => {
-                            return resolve('existing->existing (source: voice_list | dest: portal_list) has been handled before');
-                        })
-                        .catch(e => {
-                            return reject(`an error occurred while creating voice channel: ${e}`);
-                        });
-                }
-                else if (included_in_voice_list(new_channel.id, guild_object.portal_list)) { // moved from voice to voice
-                    five_min_refresher(new_channel, guild_object.portal_list, newState.guild, 5);
+      return 'existing->existing (source: voiceList | dest: portalList) has been handled before';
+    } else if (includedInVoiceList(newChannel.id, pGuild.pChannels)) {
+      // moved from voice to voice
+      await fiveMinuteRefresher(newChannel, pGuild.pChannels, newState.guild, 5);
 
-                    return resolve('existing->existing (source: voice_list | dest: voice_list)');
-                }
-                else { // moved from voice to otherefresher(new_channel, guild_object.portal_list, newState.guild, 5);
+      return 'existing->existing (source: voiceList | dest: voiceList)';
+    } else {
+      // moved from voice to other refresher(newChannel, pGuild.portalList, newState.guild, 5);
+      return 'existing->existing (source: voiceList | dest: other)';
+    }
+  } else {
+    // Joined portal channel
+    if (includedInPChannels(newChannel.id, pGuild.pChannels)) {
+      const pChannel = pGuild.pChannels.find((p) => p.id === newChannel.id);
 
-                    return resolve('existing->existing (source: voice_list | dest: other)');
-                }
-            }
-            else {
-                // Joined portal channel
-                if (included_in_portal_list(new_channel.id, guild_object.portal_list)) {
-                    const portal_object = guild_object.portal_list
-                        .find(p => p.id === new_channel.id);
+      if (!pChannel) {
+        return 'existing->existing (source: other voice | dest: portalList) / could not find portal in DB, contact Portal support';
+      }
 
-                    if (!portal_object) {
-                        return reject('existing->existing (source: other voice | dest: portal_list) / could not find portal in DB, contact Portal support');
-                    }
+      await createVoiceChannel(newState, pChannel)
+        .catch((e) => {
+          return `existing->existing (source: other voice | dest: portalList ): ${e}`;
+        });
 
-                    create_voice_channel(newState, portal_object)
-                        .then(() => {
-                            return resolve('existing->existing (source: other voice | dest: portal_list)');
-                        })
-                        .catch(e => {
-                            return reject(`existing->existing (source: other voice | dest: portal_list ): ${e}`);
-                        });
-                }
-                else if (included_in_voice_list(new_channel.id, guild_object.portal_list)) { // left created channel and joins another created
-                    five_min_refresher(new_channel, guild_object.portal_list, newState.guild, 5);
+      return 'existing->existing (source: other voice | dest: portalList)';
+    } else if (includedInVoiceList(newChannel.id, pGuild.pChannels)) {
+      // left created channel and joins another created
+      await fiveMinuteRefresher(newChannel, pGuild.pChannels, newState.guild, 5);
+      return 'existing->existing (source: other voice | dest: voiceList)';
+    }
+  }
 
-                    return resolve('existing->existing (source: other voice | dest: voice_list)');
-                }
-            }
-        }
-    });
-}
-
-module.exports = async (
-    args: { client: Client, newState: VoiceState, oldState: VoiceState }
-): Promise<string> => {
-    return new Promise((resolve, reject) => {
-        if (args.newState?.guild) {
-            const new_channel = args.newState.channel; // join channel
-            const old_channel = args.oldState.channel; // left channel
-
-            fetch_guild(args.newState?.guild.id)
-                .then(guild_object => {
-                    if (guild_object) {
-                        if (new_channel) {
-                            for (let i = 0; i < guild_object.portal_list.length; i++) {
-                                const p = guild_object.portal_list[i];
-
-                                if (p.id === new_channel.id) {
-                                    if (p.no_bots && args.newState.member?.user.bot) {
-                                        args.newState
-                                            .disconnect('voice channel does not allow bots')
-                                            .catch(e => {
-                                                return reject(`failed to kick: ${e}`);
-                                            });
-
-                                        channel_empty_check(new_channel as VoiceChannel, guild_object, args.client)
-                                            .catch(e => {
-                                                logger.error(new Error(`failed to check channel state: ${e}`));
-                                            });
-
-                                        return reject(`portal channel does not allow bots`);
-                                    }
-                                }
-
-                                for (let i = 0; i < p.voice_list.length; i++) {
-                                    const v = p.voice_list[i];
-
-                                    if (v.id === new_channel.id) {
-                                        if (v.no_bots) {
-                                            args.newState
-                                                .disconnect('voice channel does not allow bots')
-                                                .catch(e => {
-                                                    return reject(`failed to kick: ${e}`);
-                                                });
-
-                                            channel_empty_check(new_channel as VoiceChannel, guild_object, args.client)
-                                                .catch(e => {
-                                                    logger.error(new Error(`failed to check channel state: ${e}`));
-                                                });
-
-                                            return reject(`voice channel does not allow bots`);
-                                        }
-                                    }
-                                }
-                            }
-                        }
-
-                        // if (args.client.voice && args.newState.member) {
-                        //     const new_voice_connection = args.client.voice.connections
-                        //         .find((connection: VoiceConnection) =>
-                        //             !!new_channel && connection.channel.id === new_channel.id);
-
-                        //     if (new_voice_connection && !args.newState.member.user.bot) {
-                        //         client_talk(args.client, guild_object, 'user_connected');
-                        //     }
-
-                        //     const old_voice_connection = args.client.voice.connections
-                        //         .find((connection: VoiceConnection) =>
-                        //             !!old_channel && connection.channel.id === old_channel.id);
-
-                        //     if (old_voice_connection && !args.newState.member.user.bot) {
-                        //         client_talk(args.client, guild_object, 'user_disconnected');
-                        //     }
-                        // }
-
-                        if (!old_channel) {
-                            from_null(new_channel as VoiceChannel | null, guild_object, args.newState)
-                                .then(r => {
-                                    return resolve(r);
-                                })
-                                .catch(e => {
-                                    return reject(e);
-                                })
-                        } else {
-                            from_existing(old_channel as VoiceChannel, new_channel as VoiceChannel | null, args.client, guild_object, args.newState)
-                                .then(r => {
-                                    return resolve(r);
-                                })
-                                .catch(e => {
-                                    return reject(e);
-                                });
-                        }
-                    } else {
-                        return reject('could not find guild in Portal');
-                    }
-                })
-                .catch(e => {
-                    return reject(`could not find guild in Portal (${e})`);
-                });
-        } else {
-            return reject('could fnot find guild in Portal');
-        }
-    });
+  return 'existing->existing (source: other voice | dest: voiceList)';
 }
