@@ -1,862 +1,792 @@
-import { getVoiceConnection, joinVoiceChannel, VoiceConnection } from "@discordjs/voice";
-import { Client, Collection, ColorResolvable, Guild, GuildBasedChannel, GuildMember, Message, MessageEmbed, PermissionResolvable, TextBasedChannel, TextChannel, User, VoiceChannel } from "discord.js";
-import moment from "moment";
-import { createLogger, format } from "winston";
-import { VideoSearchResult } from "yt-search";
-import { GuildPrtl, MusicData } from "../types/classes/GuildPrtl.class";
-import { Field, TimeElapsed } from "../types/classes/TypesPrtl.interface";
-import { createDiscordJSAdapter } from "./adapter.library";
-// import { client_talk } from "./localisation.library";
-import { fetch_guild, fetch_guild_list, set_music_data } from "./mongo.library";
+import { getVoiceConnection, joinVoiceChannel, VoiceConnection } from '@discordjs/voice';
+import dayjs from 'dayjs';
+import duration from 'dayjs/plugin/duration';
+import relativeTime from 'dayjs/plugin/relativeTime';
+import {
+  ChatInputCommandInteraction,
+  Client,
+  Collection,
+  ColorResolvable,
+  EmbedAuthorOptions,
+  EmbedBuilder,
+  Guild,
+  GuildBasedChannel,
+  GuildMember,
+  Message,
+  PermissionResolvable,
+  TextBasedChannel,
+  TextChannel,
+  User,
+  VoiceChannel,
+} from 'discord.js';
+import { createLogger, format, transports } from 'winston';
+import { VideoSearchResult } from 'yt-search';
+import { MusicData, PGuild } from '../types/classes/PGuild.class';
+import { Field, TimeElapsed } from '../types/classes/PTypes.interface';
+import { Locale } from '../types/enums/Locales.enum';
+import { OpapGameId } from '../types/enums/OpapGames.enum';
+import { ProfanityLevel } from '../types/enums/ProfanityLevel.enum';
+import { RankSpeed } from '../types/enums/RankSpeed.enum';
+import { createDiscordJSAdapter } from './adapter.library';
+import { fetchGuild, fetchGuildList, setMusicData } from './mongo.library';
+import commandConfig from '../config.command.json';
 
-const idle_thumbnail = 'https://raw.githubusercontent.com/keybraker/' +
-	'Portal/master/src/assets/img/empty_queue.png';
+const idleThumbnail = 'https://raw.githubusercontent.com/keybraker/' + 'Portal/master/src/assets/img/emptyQueue.png';
 
 const deletedMessages = new WeakSet<Message>();
 const deletedChannel = new WeakSet<GuildBasedChannel | TextBasedChannel>();
 const deletedGuild = new WeakSet<Guild>();
 
 export function isMessageDeleted(message: Message) {
-	return deletedMessages.has(message);
+  return deletedMessages.has(message);
 }
 
 export function markMessageAsDeleted(message: Message) {
-	deletedMessages.add(message);
+  deletedMessages.add(message);
 }
 
 export function isChannelDeleted(channel: GuildBasedChannel | TextBasedChannel) {
-	return deletedChannel.has(channel);
+  return deletedChannel.has(channel);
 }
 
 export function markChannelAsDeleted(channel: GuildBasedChannel) {
-	deletedChannel.add(channel);
+  deletedChannel.add(channel);
 }
 
 export function isGuildDeleted(guild: Guild) {
-	return deletedGuild.has(guild);
+  return deletedGuild.has(guild);
 }
 
 export function markGuildAsDeleted(guild: Guild) {
-	deletedGuild.add(guild);
+  deletedGuild.add(guild);
 }
 
 export const logger = createLogger({
-	format: format.combine(
-		format.timestamp({
-			format: 'DD-MM-YY HH:mm:ss'
-		}),
-		format.errors({ stack: true }),
-		format.splat(),
-		format.json()
-	),
-	defaultMeta: { service: 'portal' },
-	// you can also add a mongo transport to store logs
-	// in the database (there is a performance penalty)
-	transports: []
+  format: format.combine(
+    format.timestamp({
+      format: 'DD-MM-YY HH:mm:ss',
+    }),
+    format.errors({ stack: true }),
+    format.splat(),
+    format.json()
+  ),
+  defaultMeta: { service: 'portal' },
+  transports: [
+    new transports.Console(),
+    new transports.File({ filename: 'error.log', level: 'error' }),
+    new transports.File({ filename: 'combined.log' }),
+  ],
 });
 
-
 export async function askForApproval(
-	message: Message, requester: GuildMember, question: string
+  interaction: ChatInputCommandInteraction,
+  requester: GuildMember,
+  question: string
 ): Promise<boolean> {
-	return new Promise((resolve, reject) => {
-		message.channel
-			.send(question)
-			.then(question_msg => {
-				let accepted = false;
-				const filter = (m: Message) => m.author.id === requester.user.id;
-				const collector = message.channel
-					.createMessageCollector({ filter, time: 10000 });
+  const questionMessage = await interaction.channel?.send(question);
 
-				collector.on('collect', (m: Message) => {
-					if (m.content === 'yes') {
-						accepted = true;
-						collector.stop();
-					}
-					else if (m.content === 'no') {
-						collector.stop();
-					}
-				});
+  if (!questionMessage) {
+    return false;
+  }
 
-				collector.on('end', async collected => {
-					for (const reply_message of collected.values()) {
-						if (isMessageDeleted(reply_message)) {
-							const deletedMessage = await reply_message
-								.delete()
-								.catch((e: any) => {
-									return reject(`failed to delete message: ${e}`);
-								});
+  let accepted = false;
+  const filter = (m: Message) => m.author.id === requester.user.id;
+  const collector = interaction.channel?.createMessageCollector({ filter, time: 10000 });
 
-							if (deletedMessage) {
-								markMessageAsDeleted(deletedMessage);
-							}
-						}
-					}
+  collector?.on('collect', (m: Message) => {
+    if (m.content === 'yes') {
+      accepted = true;
+      collector?.stop();
+    } else if (m.content === 'no') {
+      collector?.stop();
+    }
+  });
 
-					if (isMessageDeleted(question_msg)) {
-						const deletedMessage = await question_msg
-							.delete()
-							.catch((e: any) => {
-								return reject(`failed to delete message: ${e}`);
-							});
+  return new Promise((resolve) => {
+    collector?.on('end', async (collected) => {
+      for (const replyMessage of collected.values()) {
+        if (isMessageDeleted(replyMessage)) {
+          const deletedMessage = await replyMessage.delete();
 
-						if (deletedMessage) {
-							markMessageAsDeleted(deletedMessage);
-						}
-					}
+          if (deletedMessage) {
+            markMessageAsDeleted(deletedMessage);
+          }
+        }
+      }
 
-					return resolve(accepted);
-				});
-			})
-			.catch(e => {
-				return reject(e);
-			});
-	});
+      if (isMessageDeleted(questionMessage)) {
+        const deletedMessage = await questionMessage.delete();
+
+        if (deletedMessage) {
+          markMessageAsDeleted(deletedMessage);
+        }
+      }
+
+      resolve(accepted);
+    });
+  }).then(() => accepted);
 }
 
-export function getJsonFromString(
-	str: string
-): any | unknown {
-	let data = null;
+export function getJSONFromString(str: string) {
+  let data = null;
 
-	try {
-		// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-		data = JSON.parse(str);
-	}
-	catch (error) {
-		return null;
-	}
+  try {
+    data = JSON.parse(str);
+  } catch (error) {
+    return null;
+  }
 
-	// eslint-disable-next-line @typescript-eslint/no-unsafe-return
-	return data;
+  return data;
 }
 
-export function maxString(
-	abstract: string, max: number
-): string {
-	return abstract.length < max
-		? abstract
-		: abstract.substring(0, max - 3) + '...';
+export function maxString(abstract: string, max: number): string {
+  return abstract.length < max ? abstract : abstract.substring(0, max - 3) + '...';
 }
 
-export function getKeyFromEnum(
-	// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-	value: string, enumeration: any
-): string | number | undefined {
-	for (const e in enumeration) {
-		if (e === value) {
-			// eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-member-access
-			return enumeration[e];
-		}
-	}
+type enumTypes = typeof OpapGameId | typeof RankSpeed | typeof ProfanityLevel | typeof Locale;
+export function getKeyFromEnum(value: string, enumeration: enumTypes): string | number | undefined {
+  let enumerationArray;
 
-	return undefined;
+  switch (enumeration) {
+    case OpapGameId:
+      enumerationArray = Object.values(OpapGameId);
+      break;
+    case RankSpeed:
+      enumerationArray = Object.values(RankSpeed);
+      break;
+    case ProfanityLevel:
+      enumerationArray = Object.values(ProfanityLevel);
+      break;
+    case Locale:
+      enumerationArray = Object.values(Locale);
+      break;
+    default:
+      return undefined;
+  }
+
+  for (const enumerationValue of enumerationArray) {
+    if (enumerationValue === value) {
+      return enumerationValue;
+    }
+  }
+
+  return undefined;
 }
 
-export function createMusicMessage(
-	channel: TextChannel, guild_object: GuildPrtl
-): Promise<string> {
-	return new Promise((resolve, reject) => {
-		const music_message_emb = createEmded(
-			'Music Player',
-			'Type and Portal will play it !',
-			'#e60026',
-			[
-				{ emote: 'Duration', role: '-', inline: true },
-				{ emote: 'Views', role: '-', inline: true },
-				{ emote: 'Pinned', role: guild_object.music_data.pinned ? 'yes' : 'no', inline: true },
-				{ emote: 'Queue', role: 'empty', inline: false },
-				{ emote: 'Latest Action', role: '```music message created```', inline: false }
-			],
-			null,
-			null,
-			true,
-			null,
-			idle_thumbnail,
-			'https://raw.githubusercontent.com/keybraker/Portal/master/src/assets/img/music.png'
-		);
+export async function createMusicMessage(channel: TextChannel, pGuild: PGuild): Promise<string> {
+  const musicMessageEmb = createEmbed(
+    'Music Player',
+    'Type and Portal will play it !',
+    '#e60026',
+    [
+      { emote: 'Duration', role: '-', inline: true },
+      { emote: 'Views', role: '-', inline: true },
+      { emote: 'Pinned', role: pGuild.musicData.pinned ? 'yes' : 'no', inline: true },
+      { emote: 'Queue', role: 'empty', inline: false },
+      { emote: 'Latest Action', role: '```music message created```', inline: false },
+    ],
+    null,
+    null,
+    true,
+    null,
+    idleThumbnail,
+    'https://raw.githubusercontent.com/keybraker/Portal/master/src/assets/img/music.png'
+  );
 
-		channel
-			.send({ embeds: [music_message_emb] })
-			.then(sent_message => {
-				sent_message.react('▶️')
-					.catch((e: any) => {
-						return reject(`failed to set remote: ${e}`);
-					});
-				sent_message.react('⏸')
-					.catch((e: any) => {
-						return reject(`failed to set remote: ${e}`);
-					});
-				sent_message.react('⏭')
-					.catch((e: any) => {
-						return reject(`failed to set remote: ${e}`);
-					});
-				sent_message.react('📌')
-					.catch((e: any) => {
-						return reject(`failed to set remote: ${e}`);
-					});
-				sent_message.react('📄')
-					.catch((e: any) => {
-						return reject(`failed to set remote: ${e}`);
-					});
-				sent_message.react('⬇️')
-					.catch((e: any) => {
-						return reject(`failed to set remote: ${e}`);
-					});
-				sent_message.react('🧹')
-					.catch((e: any) => {
-						return reject(`failed to set remote: ${e}`);
-					});
-				sent_message.react('🚪')
-					.catch((e: any) => {
-						return reject(`failed to set remote: ${e}`);
-					});
+  const sentMessage = await channel.send({ embeds: [musicMessageEmb] })
 
-				const music_data = new MusicData(
-					channel.id,
-					sent_message.id,
-					guild_object.music_data.message_lyrics_id
-						? guild_object.music_data.message_lyrics_id
-						: 'null',
-					[],
-					false
-				);
+  if (!sentMessage) {
+    return 'failed to send message to channel';
+  }
 
-				set_music_data(guild_object.id, music_data)
-					.catch((e: any) => {
-						return reject(`failed to set music data: ${e}`);
-					});
+  const emojis = ['▶️', '⏸', '⏭', '📌', '📄', '⬇️', '🧹', '🚪'];
 
-				return resolve(sent_message.id);
-			})
-			.catch(() => {
-				return reject('failed to send message to channel');
-			});
-	});
+  for (const emoji of emojis) {
+    sentMessage.react(emoji).catch((e) => logger.error(`failed to react to message: ${e}`));
+  }
+
+  const musicData = new MusicData(
+    channel.id,
+    sentMessage.id,
+    pGuild.musicData.messageLyricsId ? pGuild.musicData.messageLyricsId : 'null',
+    [],
+    false
+  );
+
+  setMusicData(pGuild.id, musicData).catch((e) => logger.error(`failed to set music data: ${e}`));
+
+  return sentMessage.id;
 }
 
-export function createMusicLyricsMessage(
-	channel: TextChannel, guild_object: GuildPrtl, message_id: string
-): Promise<string> {
-	return new Promise((resolve, reject) => {
-		const music_lyrics_message_emb = createEmded(
-			'Lyrics 📄',
-			'',
-			'#e60026',
-			null,
-			null,
-			null,
-			false,
-			null,
-			null
-		);
+export async function createMusicLyricsMessage(channel: TextChannel, pGuild: PGuild, messageId: string): Promise<string | undefined> {
+  const musicLyricsMessageEmb = createEmbed('Lyrics 📄', '', '#e60026', null, null, null, false, null, null);
 
-		channel
-			.send({ embeds: [music_lyrics_message_emb] })
-			.then(sent_message_lyrics => {
-				const music_data = new MusicData(
-					channel.id,
-					message_id,
-					sent_message_lyrics.id,
-					[],
-					false
-				);
+  const sentMessageLyrics = await channel.send({ embeds: [musicLyricsMessageEmb] });
 
-				set_music_data(guild_object.id, music_data)
-					.catch((e: any) => {
-						return reject(`failed to set music data: ${e}`);
-					});
+  if (!sentMessageLyrics) {
+    return undefined;
+  }
 
-				return resolve(sent_message_lyrics.id);
-			})
-			.catch(e => {
-				return reject(`failed to send message to channel: ${e}`);
-			});
-	});
+  const musicData = new MusicData(channel.id, messageId, sentMessageLyrics.id, [], false);
+  await setMusicData(pGuild.id, musicData);
+
+  return sentMessageLyrics.id;
 }
 
-export function updateMusicMessage(
-	guild: Guild, guild_object: GuildPrtl, yts: VideoSearchResult | undefined,
-	status: string, animated = true
+export async function updateMusicMessage(
+  guild: Guild,
+  pGuild: PGuild,
+  yts: VideoSearchResult | undefined,
+  status: string,
+  animated = true
 ): Promise<boolean> {
-	return new Promise((resolve, reject) => {
-		const guild_channel: GuildBasedChannel | undefined = guild.channels.cache
-			.find(c => c.id === guild_object.music_data.channel_id);
+  const guildChannel: GuildBasedChannel | undefined = guild.channels.cache.find(
+    (c) => c.id === pGuild.musicData.channelId
+  );
 
-		if (!guild_channel) {
-			return reject(`could not fetch channel`);
-		}
+  if (!guildChannel) {
+    return false
+  }
 
-		const channel: TextChannel = <TextChannel>guild_channel;
+  const channel: TextChannel = <TextChannel>guildChannel;
 
-		if (!channel || !guild_object.music_data.message_id) {
-			return reject(`could not find channel`);
-		}
+  if (!channel || !pGuild.musicData.messageId) {
+    return false
+  }
 
-		const music_queue = guild_object.music_queue ?
-			guild_object.music_queue.length > 1
-				? guild_object.music_queue
-					.map((v, i) => {
-						if (i !== 0 && i < 6) {
-							return (`${i}. ${maxString(v.title, 61)}`);
-						} else if (i === 6) {
-							return `_...${guild_object.music_queue.length - 6} more_`;
-						}
-					})
-					.filter(v => !!v)
-					.join('\n')
-				: 'empty'
-			: 'empty';
+  const musicQueue = pGuild.musicQueue
+    ? pGuild.musicQueue.length > 1
+      ? pGuild.musicQueue
+        .map((v, i) => {
+          if (i !== 0 && i < 6) {
+            return `${i}. ${maxString(v.title, 61)}`;
+          } else if (i === 6) {
+            return `...${pGuild.musicQueue.length - 6} more`;
+          }
+        })
+        .filter((v) => !!v)
+        .join('\n')
+      : 'empty'
+    : 'empty';
 
-		const music_message_emb = createEmded(
-			yts ? yts.title : 'Music Player',
-			yts ? yts.url : 'Type and Portal will play it !',
-			'#e60026',
-			[
-				{ emote: 'Duration', role: yts ? yts.timestamp : '-', inline: true },
-				{ emote: 'Views', role: (yts ? yts.timestamp : 0) === 0 ? '-' : yts ? yts.views : '-', inline: true },
-				{ emote: 'Pinned', role: guild_object.music_data.pinned ? 'yes' : 'no', inline: true },
-				// { emote: null, role: null, inline: true },
-				{ emote: 'Queue', role: music_queue, inline: false },
-				// { emote: null, role: null, inline: true },
-				{ emote: 'Latest Action', role: '```' + status + '```', inline: false }
-			],
-			null,
-			null,
-			true,
-			null,
-			yts ? yts.thumbnail : idle_thumbnail,
-			animated
-				? 'https://raw.githubusercontent.com/keybraker/Portal/master/src/assets/img/music.gif'
-				: 'https://raw.githubusercontent.com/keybraker/Portal/master/src/assets/img/music.png'
-		);
+  const musicMessageEmb = createEmbed(
+    yts ? yts.title : 'Music Player',
+    yts ? yts.url : 'Type and Portal will play it !',
+    '#e60026',
+    [
+      { emote: 'Duration', role: yts ? yts.timestamp : '-', inline: true },
+      { emote: 'Views', role: (yts ? yts.timestamp : 0) === 0 ? '-' : yts ? yts.views : '-', inline: true },
+      { emote: 'Pinned', role: pGuild.musicData.pinned ? 'yes' : 'no', inline: true },
+      // { emote: null, role: null, inline: true },
+      { emote: 'Queue', role: musicQueue, inline: false },
+      // { emote: null, role: null, inline: true },
+      { emote: 'Latest Action', role: '```' + status + '```', inline: false },
+    ],
+    null,
+    null,
+    true,
+    null,
+    yts ? yts.thumbnail : idleThumbnail,
+    animated
+      ? 'https://raw.githubusercontent.com/keybraker/Portal/master/src/assets/img/music.gif'
+      : 'https://raw.githubusercontent.com/keybraker/Portal/master/src/assets/img/music.png'
+  );
 
-		if (guild_object.music_data.message_id) {
-			if (channel) {
-				channel.messages
-					.fetch(guild_object.music_data.message_id)
-					.then((message: Message) => {
-						message
-							.edit({ embeds: [music_message_emb] })
-							.then(() => {
-								return resolve(true);
-							})
-							.catch(e => {
-								return reject(`failed to edit messages: ${e}`);
-							});
-					})
-					.catch(e => {
-						return reject(`failed to fetch messages: ${e}`);
-					});
-			}
-		}
-	});
+  if (!pGuild.musicData.messageId || !channel) {
+    return false;
+  }
+
+  const message = await channel.messages.fetch(pGuild.musicData.messageId);
+
+  if (!message) {
+    return false;
+  }
+
+  const messageEdit = await message.edit({ embeds: [musicMessageEmb] });
+
+  return !!messageEdit;
 }
 
-export function updateMusicLyricsMessage(
-	guild: Guild, guild_object: GuildPrtl, lyrics: string, url?: string
-): Promise<boolean> {
-	return new Promise((resolve, reject) => {
-		const guild_channel: GuildBasedChannel | undefined = guild.channels.cache
-			.find(c => c.id === guild_object.music_data.channel_id);
+export async function updateMusicLyricsMessage(guild: Guild, pGuild: PGuild, lyrics: string, url?: string): Promise<boolean> {
+  const guildChannel = guild.channels.cache.find((c) => c.id === pGuild.musicData.channelId);
 
-		if (!guild_channel) {
-			return reject(`could not fetch channel`);
-		}
+  if (!guildChannel) {
+    return false;
+  }
 
-		const channel: TextChannel = <TextChannel>guild_channel;
+  const channel: TextChannel = <TextChannel>guildChannel;
 
-		if (!channel || !guild_object.music_data.message_id) {
-			return reject(`could not find channel`);
-		}
+  if (!channel || !pGuild.musicData.messageId) {
+    return false;
+  }
 
-		const music_message_emb = createEmded(
-			`Lyrics 📄 ${url ? `at ${url}` : ''}`,
-			maxString(lyrics, 2000),
-			'#e60026',
-			null,
-			null,
-			null,
-			false,
-			null,
-			null
-		);
+  const musicMessageEmb = createEmbed(
+    `Lyrics 📄 ${url ? `at ${url}` : ''}`,
+    maxString(lyrics, 2000),
+    '#e60026',
+    null,
+    null,
+    null,
+    false,
+    null,
+    null
+  );
 
-		if (guild_object.music_data.message_lyrics_id) {
-			if (channel) {
-				channel.messages
-					.fetch(guild_object.music_data.message_lyrics_id)
-					.then((message: Message) => {
-						message.edit({ embeds: [music_message_emb] })
-							.then(() => {
-								return resolve(true);
-							})
-							.catch(e => {
-								return reject(`failed to edit messages: ${e}`);
-							});
-					})
-					.catch(e => {
-						return reject(`failed to fetch messages: ${e}`);
-					});
-			}
-		}
-	});
+  if (!pGuild.musicData.messageLyricsId || !channel) {
+    return false;
+  }
+
+  const fetchedMessage = await channel.messages.fetch(pGuild.musicData.messageLyricsId);
+
+  if (!fetchedMessage) {
+    return false;
+  }
+
+  const editedMessage = await fetchedMessage.edit({ embeds: [musicMessageEmb] });
+
+  return !!editedMessage;
 }
 
 export async function joinUserVoiceChannelByReaction(
-	guild: Guild, client: Client, guild_object: GuildPrtl, user: User, announce_entrance: boolean
+  guild: Guild,
+  client: Client,
+  pGuild: PGuild,
+  user: User
+  // announceEntrance: boolean
 ): Promise<VoiceConnection> {
-	const guildMembers = await guild.members.fetch();
+  const guildMembers = await guild.members.fetch();
 
-	if (!guildMembers) {
-		return Promise.reject(`could not fetch members`);
-	}
+  if (!guildMembers) {
+    return Promise.reject('could not fetch members');
+  }
 
-	const member = guildMembers.find(m => !m.user?.bot && m.id === user.id);
+  const member = guildMembers.find((guildMember) => !guildMember.user?.bot && guildMember.id === user.id);
 
-	if (!member) {
-		return Promise.reject(`could not find member`);
-	}
+  if (!member) {
+    return Promise.reject('could not find member');
+  }
 
-	if (!member.voice) {
-		return Promise.reject('you must be connected to a voice channel');
-	}
+  if (!member.voice) {
+    return Promise.reject('you must be connected to a voice channel');
+  }
 
-	if (!member.voice.channel) {
-		return Promise.reject('you must be connected to a voice channel');
-	}
+  if (!member.voice.channel) {
+    return Promise.reject('you must be connected to a voice channel');
+  }
 
-	let voiceConnection = await getVoiceConnection(member.voice.channel.id);
-	if (voiceConnection && member.voice.channel.guild.me?.voice.channelId === member.voice.channel?.id) {
-		member.voice.channel.guild.me?.voice.setDeaf();
-	} else {
-		voiceConnection = joinVoiceChannel({
-			channelId: member.voice.channel.id,
-			guildId: member.voice.channel.guild.id,
-			adapterCreator: createDiscordJSAdapter(member.voice.channel as VoiceChannel),
-		});
+  let voiceConnection = await getVoiceConnection(member.voice.channel.id);
+  const clientVoiceState = member.voice.channel.guild.voiceStates.cache.get(member.client.user.id);
 
-		if (!voiceConnection) {
-			return Promise.reject('could not join voice channel');
-		}
+  if (voiceConnection && clientVoiceState?.channelId === member.voice.channel?.id) {
+    // member.voice.channel.guild.me?.voice.setDeaf();
+    clientVoiceState.setDeaf(true);
+  } else {
+    voiceConnection = joinVoiceChannel({
+      channelId: member.voice.channel.id,
+      guildId: member.voice.channel.guild.id,
+      adapterCreator: createDiscordJSAdapter(member.voice.channel as VoiceChannel),
+    });
 
-		member.voice.channel.guild.me?.voice.setDeaf();
-	}
+    if (!voiceConnection) {
+      return Promise.reject('could not join voice channel');
+    }
 
-	return voiceConnection;
+    // member.voice.channel.guild.me?.voice.setDeaf();
+    if (clientVoiceState) {
+      clientVoiceState.setDeaf(true);
+    }
+  }
+
+  return voiceConnection;
 }
-
-export async function joinUserVoiceChannelByMessage(
-	client: Client, message: Message, guild_object: GuildPrtl, join = false
+export async function joinUserVoiceChannelByInteraction(
+  client: Client,
+  interaction: ChatInputCommandInteraction,
+  pGuild: PGuild
+  // join = false
 ): Promise<VoiceConnection> {
-	if (!message.member) {
-		return Promise.reject('user could not be fetched for message');
-	}
+  const member = interaction.member as GuildMember;
 
-	if (!message.member.voice) {
-		return Promise.reject('voice could not be fetched for member');
-	}
+  if (!member) {
+    return Promise.reject('user could not be fetched for message');
+  }
 
-	if (!message.member.voice.channel) {
-		return Promise.reject('you aren\'t in a channel');
-	}
+  if (!member.voice) {
+    return Promise.reject('voice could not be fetched for member');
+  }
 
-	if (!message.guild) {
-		return Promise.reject('guild could not be fetched for message');
-	}
+  if (!member.voice.channel) {
+    return Promise.reject('you aren\'t in a channel');
+  }
 
-	if (!message.guild.voiceAdapterCreator) {
-		return Promise.reject('voiceAdapterCreator could not be fetched for guild');
-	}
+  if (!interaction.guild) {
+    return Promise.reject('guild could not be fetched for message');
+  }
 
-	if (!guild_object) {
-		return Promise.reject('could not find guild of message');
-	}
+  if (!interaction.guild.voiceAdapterCreator) {
+    return Promise.reject('voiceAdapterCreator could not be fetched for guild');
+  }
 
-	if (!client.voice) {
-		return Promise.reject('could not fetch portal\'s voice connections');
-	}
+  if (!pGuild) {
+    return Promise.reject('could not find guild of message');
+  }
 
-	let voiceConnection = await getVoiceConnection(message.member.voice.channel.id);
-	if (voiceConnection && message.guild.me?.voice.channelId === message.member.voice.channel?.id) {
-		message.guild.me?.voice.setDeaf();
-	} else {
-		voiceConnection = await joinVoiceChannel({
-			channelId: message.member.voice.channel.id,
-			guildId: message.guild.id,
-			adapterCreator: createDiscordJSAdapter(message.member.voice.channel as VoiceChannel),
-		});
+  if (!client.voice) {
+    return Promise.reject('could not fetch portal\'s voice connections');
+  }
 
-		if (!voiceConnection) {
-			return Promise.reject('could not join voice channel');
-		}
+  let voiceConnection = await getVoiceConnection(member.voice.channel.id);
+  const clientVoiceState = interaction.guild.voiceStates.cache.get(interaction.guild.client.user.id);
 
-		message.guild.me?.voice.setDeaf();
-	}
+  if (voiceConnection && clientVoiceState?.channelId === member.voice.channel?.id) {
+    // message.guild.me?.voice.setDeaf();
+    clientVoiceState.setDeaf(true);
+  } else {
+    voiceConnection = await joinVoiceChannel({
+      channelId: member.voice.channel.id,
+      guildId: interaction.guild.id,
+      adapterCreator: createDiscordJSAdapter(member.voice.channel as VoiceChannel),
+    });
 
-	return voiceConnection;
+    if (!voiceConnection) {
+      return Promise.reject('could not join voice channel');
+    }
+
+    // message.guild.me?.voice.setDeaf();
+    if (clientVoiceState) {
+      clientVoiceState.setDeaf(true);
+    }
+  }
+
+  return voiceConnection;
 }
 
-export function createEmded(
-	title: string | null | undefined,
-	description: string | null | undefined,
-	colour: ColorResolvable | null | undefined,
-	field_array: Field[] | null,
-	thumbnail: string | null | undefined,
-	member: GuildMember | null | undefined,
-	from_bot: boolean | null | undefined,
-	url: string | null | undefined,
-	image: string | null | undefined,
-	custom_gif?: string,
-	author?: { name: string, icon: string },
-	footer?: string,
-): MessageEmbed {
-	const portal_icon_url: string = 'https://raw.githubusercontent.com/keybraker' +
-		'/Portal/master/src/assets/img/portal_logo_spinr.gif';
+export function createEmbed(
+  title: string | null | undefined,
+  description: string | null | undefined,
+  colour: ColorResolvable | null | undefined,
+  fieldArray: Field[] | null,
+  thumbnail: string | null | undefined,
+  member: GuildMember | null | undefined,
+  fromBot: boolean | null | undefined,
+  url: string | null | undefined,
+  image: string | null | undefined,
+  customGif?: string,
+  author?: { name: string; icon: string },
+  footer?: string
+): EmbedBuilder {
+  const portalIconUrl: string =
+    'https://raw.githubusercontent.com/keybraker' + '/Portal/master/src/assets/img/portal_logo_spinr.gif';
 
-	const rich_message: MessageEmbed = new MessageEmbed();
+  const richMessage = new EmbedBuilder();
 
-	if (title) rich_message.setTitle(title);
-	if (url) rich_message.setURL(url);
-	if (colour) rich_message.setColor(colour);
-	if (description) rich_message.setDescription(description);
-	if (footer) rich_message.setFooter({ text: footer });
-	if (from_bot) rich_message.setFooter({
-		text: footer
-			? footer
-			: 'Portal',
-		iconURL: custom_gif
-			? custom_gif
-			: portal_icon_url
-	}).setTimestamp();
-	if (thumbnail) rich_message.setThumbnail(thumbnail);
-	if (image) rich_message.setImage(image);
-	if (author) rich_message.setAuthor({
-		name: author.name,
-		iconURL: author.icon
-		// url: 'https://discord.js.org'
-	});
-	if (field_array) {
-		field_array.forEach(row => {
-			rich_message
-				.addField(
-					row.emote === '' || !row.emote
-						? `\u200b`
-						: `__${row.emote}__`,
-					row.role === '' || !row.role
-						? `\u200b`
-						: `${row.role}`,
-					row.inline
-				);
-		});
-	}
-	if (member && !author) {
-		const url = member.user.avatarURL()
-			? member.user.avatarURL()
-			: undefined;
+  if (title) richMessage.setTitle(title);
+  if (url) richMessage.setURL(url);
+  if (colour) richMessage.setColor(colour);
+  if (description) richMessage.setDescription(description);
+  if (footer) richMessage.setFooter({ text: footer });
+  if (fromBot)
+    richMessage
+      .setFooter({
+        text: footer ?? 'Portal',
+        iconURL: customGif ?? portalIconUrl,
+      })
+      .setTimestamp();
+  if (thumbnail) richMessage.setThumbnail(thumbnail);
+  if (image) richMessage.setImage(image);
+  if (author)
+    richMessage.setAuthor({
+      name: author.name,
+      iconURL: author.icon,
+      // url: 'https://discord.js.org'
+    });
+  if (fieldArray) {
+    fieldArray.forEach((row) => {
+      richMessage.addFields({
+        name: row.emote === '' || !row.emote ? '\u200b' : `${row.emote}`,
+        value: row.role === '' || !row.role ? '\u200b' : `${row.role}`,
+        inline: row.inline,
+      });
+    });
+  }
+  if (member && !author) {
+    const url = member.user.avatarURL() ? member.user.avatarURL() : undefined;
 
-		rich_message.setAuthor(
-			member.displayName,
-			url
-				? url
-				: undefined,
-			undefined
-		);
-	}
+    richMessage.setAuthor({
+      name: member.displayName,
+      value: url ? url.toString() : undefined,
+      inline: undefined,
+    } as EmbedAuthorOptions);
+  }
 
-	return rich_message;
+  return richMessage;
 }
 
-export function isUserAuthorised(
-	member: GuildMember
-): boolean {
-	const administrator: PermissionResolvable = 'ADMINISTRATOR';
+export function isUserAuthorised(member: GuildMember): boolean {
+  const administrator: PermissionResolvable = 'Administrator';
 
-	if (member.permissions.has(administrator, true)) {
-		return true;
-	}
+  if (member.permissions.has(administrator, true)) {
+    return true;
+  }
 
-	if (member.roles.cache) {
-		return member.roles.cache.some(r =>
-			r.name.toLowerCase() === 'p.admin');
-	}
+  if (member.roles.cache) {
+    return member.roles.cache.some((r) => r.name.toLowerCase() === 'p.admin');
+  }
 
-	return false;
+  return false;
 }
 
-export function isUserDj(
-	member: GuildMember
-): boolean {
-	if (member.roles.cache) {
-		return member.roles.cache.some(r =>
-			r.name.toLowerCase() === 'p.dj');
-	}
+export function isUserDj(member: GuildMember): boolean {
+  if (member.roles.cache) {
+    return member.roles.cache.some((r) => r.name.toLowerCase() === 'p.dj');
+  }
 
-	return false;
+  return false;
 }
 
-export function isUserIgnored(
-	member: GuildMember
-): boolean {
-	return member.roles.cache.some(r =>
-		r.name.toLowerCase() === 'p.ignore');
+export function isUserIgnored(member: GuildMember): boolean {
+  return member.roles.cache.some((r) => r.name.toLowerCase() === 'p.ignore');
 }
 
-export function isMod(
-	member: GuildMember | null
-): boolean {
-	if (member && member.roles.cache) {
-		return member.roles.cache.some(r =>
-			r.name.toLowerCase() === 'p.mod');
-	}
+export function isMod(member: GuildMember | null): boolean {
+  if (member && member.roles.cache) {
+    return member.roles.cache.some((r) => r.name.toLowerCase() === 'p.mod');
+  }
 
-	return false;
+  return false;
 }
 
-export function isWhitelist(
-	member: GuildMember | null
-): boolean {
-	if (member && member.roles.cache) {
-		return member.roles.cache.some(r =>
-			r.name.toLowerCase() === 'p.whitelist');
-	}
+export function isWhitelist(member: GuildMember | null): boolean {
+  if (member && member.roles.cache) {
+    return member.roles.cache.some((r) => r.name.toLowerCase() === 'p.whitelist');
+  }
 
-	return false;
+  return false;
 }
 
-export function messageHelp(
-	type: string, argument: string, info = ''
-): string {
-	if (info !== '') info += '\n';
-	return `${info} get help by typing \`./help ${argument}\`\n` +
-		`*https://portal-bot.xyz/docs/${type}/detailed/${argument}*`;
+export function messageHelp(type: string, argument: string, info = ''): string {
+  if (info !== '') info += '\n';
+  return (
+    `${info} get help by typing \`./help ${argument}\`\n` + `*https://portal-bot.xyz/docs/${type}/detailed/${argument}*`
+  );
 }
 
 export async function messageReply(
-	status: boolean,
-	message: Message,
-	replyString: string,
-	deleteSource = false,
-	deleteReply = false,
-	emotePass = '✔️',
-	emoteFail = '❌'
+  status: boolean,
+  message: Message,
+  replyString: string,
+  deleteSource = false,
+  deleteReply = false,
+  emotePass = '✔️',
+  emoteFail = '❌'
 ): Promise<boolean> {
-	if (!message) {
-		return Promise.reject(`failed to find message`);
-	}
+  if (!message) {
+    return Promise.reject('failed to find message');
+  }
 
-	if (!isChannelDeleted(message.channel) && replyString !== null && replyString !== '') {
-		const sentMessage = await message.reply(replyString)
-			.catch(e => { return Promise.reject(`failed to send message: ${e}`); });
+  if (!isChannelDeleted(message.channel) && replyString !== null && replyString !== '') {
+    const sentMessage = await message.reply(replyString).catch((e) => {
+      return Promise.reject(`failed to send message: ${e}`);
+    });
 
-		if (!sentMessage) {
-			return Promise.reject(`failed to send message`);
-		}
+    if (!sentMessage) {
+      return Promise.reject('failed to send message');
+    }
 
-		if (deleteReply) {
-			const delay = (process.env.DELETE_DELAY as unknown as number) * 1000;
-			setTimeout(async () => {
-				if (isMessageDeleted(sentMessage)) {
-					const deletedMessage = await sentMessage
-						.delete()
-						.catch((e: any) => {
-							return Promise.reject(`failed to delete message: ${e}`);
-						});
+    if (deleteReply) {
+      const delay = (process.env.DELETEDELAY as unknown as number) * 1000;
+      setTimeout(async () => {
+        if (isMessageDeleted(sentMessage)) {
+          const deletedMessage = await sentMessage.delete().catch((e) => {
+            return Promise.reject(`failed to delete message: ${e}`);
+          });
 
-					if (deletedMessage) {
-						markMessageAsDeleted(deletedMessage);
-					}
-				}
-			}, delay);
-		}
-	}
+          if (deletedMessage) {
+            markMessageAsDeleted(deletedMessage);
+          }
+        }
+      }, delay);
+    }
+  }
 
-	if (deleteSource) {
-		const reaction = await message.react(status ? emotePass : emoteFail)
-			.catch(e => { return Promise.reject(`failed to react to message: ${e}`); });
+  if (deleteSource) {
+    const reaction = await message.react(status ? emotePass : emoteFail).catch((e) => {
+      return Promise.reject(`failed to react to message: ${e}`);
+    });
 
-		if (!reaction) {
-			return Promise.reject(`failed to react to message`);
-		}
+    if (!reaction) {
+      return Promise.reject('failed to react to message');
+    }
 
-		const delay = (process.env.DELETE_DELAY as unknown as number) * 1000;
-		setTimeout(async () => {
-			if (isMessageDeleted(message)) {
-				const deletedMessage = await message
-					.delete()
-					.catch((e: any) => {
-						return Promise.reject(`failed to delete message: ${e}`);
-					});
+    const delay = (process.env.DELETEDELAY as unknown as number) * 1000;
+    setTimeout(async () => {
+      if (isMessageDeleted(message)) {
+        const deletedMessage = await message.delete().catch((e) => {
+          return Promise.reject(`failed to delete message: ${e}`);
+        });
 
-				if (deletedMessage) {
-					markMessageAsDeleted(deletedMessage);
-				}
-			}
-		}, delay);
+        if (deletedMessage) {
+          markMessageAsDeleted(deletedMessage);
+        }
+      }
+    }, delay);
 
-		return true;
-	}
+    return true;
+  }
 
-	return false;
+  return false;
 }
 
-export function isUrl(
-	potential_url: string
-): boolean {
-	const pattern = new RegExp('^(https?:\\/\\/)?' + // protocol
-		'((([a-z\\d]([a-z\\d-]*[a-z\\d])*)\\.)+[a-z]{2,}|' + // domain name
-		'((\\d{1,3}\\.){3}\\d{1,3}))' + // OR ip (v4) address
-		'(\\:\\d+)?(\\/[-a-z\\d%_.~+]*)*' + // port and path
-		'(\\?[;&a-z\\d%_.~+=-]*)?' + // query string
-		'(\\#[-a-z\\d_]*)?$', 'i'); // fragment locator
+export function isUrl(potentialURL: string): boolean {
+  const pattern = new RegExp(
+    '^(https?:\\/\\/)?' + // protocol
+    '((([a-z\\d]([a-z\\d-]*[a-z\\d])*)\\.)+[a-z]{2,}|' + // domain name
+    '((\\d{1,3}\\.){3}\\d{1,3}))' + // OR ip (v4) address
+    '(\\:\\d+)?(\\/[-a-z\\d%.~+]*)*' + // port and path
+    '(\\?[;&a-z\\d%.~+=-]*)?' + // query string
+    '(\\#[-a-z\\d]*)?$',
+    'i'
+  ); // fragment locator
 
-	return pattern.test(potential_url);
+  return pattern.test(potentialURL);
 }
 
-export function pad(
-	num: number
-): string {
-	return num.toString().length >= 2
-		? `${num}`
-		: `0${num}`;
+export function pad(num: number): string {
+  return num.toString().length >= 2 ? `${num}` : `0${num}`;
 }
 
-export function timeElapsed(
-	timestamp: Date | number, timeout: number
-): TimeElapsed {
-	const timeout_time = timeout * 60 * 1000;
-	const el = moment
-		.duration(moment()
-			.diff(
-				moment(typeof timestamp === 'number'
-					? timestamp
-					: timestamp.getTime()
-				)
-			)
-		);
+dayjs.extend(duration);
+dayjs.extend(relativeTime);
 
-	const timeout_min = moment(timeout_time).minutes();
-	const timeout_sec = moment(timeout_time).seconds();
-	const remaining_hrs = el.hours();
-	const remaining_min = el.minutes();
-	const remaining_sec = el.seconds();
+export function timeElapsed(timestamp: Date | number, timeout: number): TimeElapsed {
+  const timeoutTime = timeout * 60 * 1000;
+  const el = dayjs.duration(dayjs().diff(dayjs(typeof timestamp === 'number' ? timestamp : timestamp.getTime())));
 
-	return { timeout_min, timeout_sec, remaining_hrs, remaining_min, remaining_sec }
+  const timeoutMin = dayjs.duration(timeoutTime).minutes();
+  const timeoutSec = dayjs.duration(timeoutTime).seconds();
+  const remainingHrs = el.hours();
+  const remainingMin = el.minutes();
+  const remainingSec = el.seconds();
+
+  return { timeoutMin, timeoutSec, remainingHrs, remainingMin, remainingSec };
 }
 
 // must get updated
-export function removeDeletedChannels(
-	guild: Guild
-): Promise<boolean> {
-	return new Promise((resolve) => {
-		fetch_guild(guild.id)
-			.then(guild_object => {
-				if (guild_object) {
-					guild_object.portal_list.forEach((p, index_p) => {
-						if (!guild.channels.cache.some(c => c.id === p.id)) {
-							guild_object.portal_list.splice(index_p, 1);
-						}
-						p.voice_list.forEach((v, index_v) => {
-							if (!guild.channels.cache.some(c => c.id === v.id)) {
-								p.voice_list.splice(index_v, 1);
-							}
-						});
-					});
+export function removeDeletedChannels(guild: Guild): Promise<boolean> {
+  return new Promise((resolve) => {
+    fetchGuild(guild.id)
+      .then((pGuild) => {
+        if (pGuild) {
+          pGuild.pChannels.forEach((p, indexP) => {
+            if (!guild.channels.cache.some((c) => c.id === p.id)) {
+              pGuild.pChannels.splice(indexP, 1);
+            }
+            p.pVoiceChannels.forEach((v, indexV) => {
+              if (!guild.channels.cache.some((c) => c.id === v.id)) {
+                p.pVoiceChannels.splice(indexV, 1);
+              }
+            });
+          });
 
-					guild_object.url_list.some((u_id, index_u) => {
-						if (!guild.channels.cache.some(c => c.id === u_id)) {
-							guild_object.url_list.splice(index_u, 1);
-							return true;
-						}
+          pGuild.pURLs.some((uId, indexU) => {
+            if (!guild.channels.cache.some((c) => c.id === uId)) {
+              pGuild.pURLs.splice(indexU, 1);
+              return true;
+            }
 
-						return false;
-					});
+            return false;
+          });
 
-					guild_object.role_list.forEach((r, index_r) => {
-						!guild.channels.cache.some(c => {
-							if (c instanceof TextChannel) {
-								let found = false;
-								c.messages
-									.fetch(r.message_id)
-									.then(() => {
-										// clear from emotes leave only those from portal
-										found = true;
-									})
-									.catch(() => {
-										guild_object.role_list.splice(index_r, 1);
-									});
+          pGuild.pRoles.forEach((r, indexR) => {
+            !guild.channels.cache.some((c) => {
+              if (c instanceof TextChannel) {
+                let found = false;
+                c.messages
+                  .fetch(r.messageId)
+                  .then(() => {
+                    // clear from emotes leave only those from portal
+                    found = true;
+                  })
+                  .catch(() => {
+                    pGuild.pRoles.splice(indexR, 1);
+                  });
 
-								return found;
-							}
+                return found;
+              }
 
-							return false;
-						});
-					});
+              return false;
+            });
+          });
 
-					guild_object.member_list.forEach((m, index_m) => {
-						if (!guild.members.cache.some(m => m.id === m.id)) {
-							guild_object.url_list.splice(index_m, 1);
-						}
-					});
+          pGuild.pMembers.forEach((m, indexM) => {
+            if (!guild.members.cache.some((m) => m.id === m.id)) {
+              pGuild.pURLs.splice(indexM, 1);
+            }
+          });
 
-					if (!guild.channels.cache.some(c => c.id === guild_object.music_data.channel_id)) {
-						guild_object.music_data.channel_id = undefined;
-						guild_object.music_data.message_id = undefined;
-						guild_object.music_data.votes = undefined;
-					}
+          if (!guild.channels.cache.some((c) => c.id === pGuild.musicData.channelId)) {
+            pGuild.musicData.channelId = undefined;
+            pGuild.musicData.messageId = undefined;
+            pGuild.musicData.votes = undefined;
+          }
 
-					if (!guild.channels.cache.some(c => c.id === guild_object.announcement)) {
-						guild_object.announcement = null;
-					}
+          if (!guild.channels.cache.some((c) => c.id === pGuild.announcement)) {
+            pGuild.announcement = null;
+          }
 
-					return resolve(true);
-				}
+          return resolve(true);
+        }
 
-				return resolve(false);
-			})
-			.catch(() => {
-				return resolve(false);
-			});
-	});
+        return resolve(false);
+      })
+      .catch(() => {
+        return resolve(false);
+      });
+  });
 }
 
 // must get updated
-export async function removeEmptyVoiceChannels(
-	guild: Guild
-): Promise<boolean> {
-	const guild_list = await fetch_guild_list();
-	if (!guild_list) {
-		return false;
-	}
+export async function removeEmptyVoiceChannels(guild: Guild): Promise<boolean> {
+  const guildList = await fetchGuildList({});
+  if (!guildList) {
+    return false;
+  }
 
-	if (guild_list?.length === 0) {
-		return true;
-	}
+  if (guildList?.length === 0) {
+    return true;
+  }
 
-	guild.channels.cache.forEach(channel => {
-		guild_list.some(g =>
-			g.portal_list.some(p =>
-				p.voice_list.some((v, index) => {
-					if (v.id === channel.id && (<Collection<string, GuildMember>>channel.members).size === 0) {
-						channel
-							.delete()
-							.then(() => {
-								p.voice_list.splice(index, 1);
-								logger.log({
-									level: 'info', type: 'none', message: `deleted empty channel: ${channel.name} ` +
-										`(${channel.id}) from ${channel.guild.name}`
-								});
-							})
-							.catch(e => {
-								logger.log({ level: 'error', type: 'none', message: `failed to send message: ${e}` });
-							});
-						return true;
-					}
-					return false
-				})
-			)
-		);
-	});
+  guild.channels.cache.forEach((channel) => {
+    guildList.some((g) =>
+      g.pChannels.some((p) =>
+        p.pVoiceChannels.some((v, index) => {
+          if (v.id === channel.id && (<Collection<string, GuildMember>>channel.members).size === 0) {
+            channel
+              .delete()
+              .then(() => {
+                p.pVoiceChannels.splice(index, 1);
+                logger.log({
+                  level: 'info',
+                  type: 'none',
+                  message: `deleted empty channel: ${channel.name} ` + `(${channel.id}) from ${channel.guild.name}`,
+                });
+              })
+              .catch((e) => {
+                logger.log({ level: 'error', type: 'none', message: `failed to send message: ${e}` });
+              });
+            return true;
+          }
+          return false;
+        })
+      )
+    );
+  });
 
-	return true;
+  return true;
+}
+
+export function commandDescriptionByNameAndAuthenticationLevel(commandName: string, authenticationNeeded: boolean) {
+  return authenticationNeeded
+    ? commandConfig[1].commands.find((command) => command.name === commandName)?.description ?? commandName
+    : commandConfig[0].commands.find((command) => command.name === commandName)?.description ?? commandName;
 }
