@@ -3,48 +3,43 @@ import https, { RequestOptions } from 'https';
 import fetch from 'node-fetch';
 import { URL } from 'url';
 
-export async function https_fetch(
-	options: string | RequestOptions | URL
-): Promise<Buffer> {
-	return new Promise((resolve, reject) => {
-		const req = https.request(options, function (res) {
-			const chunks: Uint8Array[] = [];
+export async function httpsFetch(options: string | RequestOptions | URL): Promise<Buffer> {
+  return new Promise((resolve, reject) => {
+    const req = https.request(options, function (res) {
+      const chunks: Uint8Array[] = [];
 
-			res.on('data', function (chunk: Uint8Array) {
-				chunks.push(chunk);
-			});
+      res.on('data', function (chunk: Uint8Array) {
+        chunks.push(chunk);
+      });
 
-			res.on('end', function () { // (chunk: any) {
-				return resolve(Buffer.concat(chunks));
-			});
+      res.on('end', function () {
+        // (chunk) {
+        return resolve(Buffer.concat(chunks));
+      });
 
-			res.on('error', function () { // (error) {
-				return reject(false);
-			});
-		});
+      res.on('error', function () {
+        // (error) {
+        return reject(false);
+      });
+    });
 
-		req.end();
-	});
+    req.end();
+  });
 }
 
-export async function scrape_lyrics(
-	url: string
-): Promise<string> {
-	return new Promise((resolve, reject) => {
-		fetch(url)
-			.then((response: any) => {
-				response
-					.text()
-					.then((text: any) => {
-						const $ = cheerio.load(text);
-						return resolve($('.lyrics').text().trim());
-					})
-					.catch((e: any) => {
-						return reject(e);
-					});
-			})
-			.catch((e: any) => {
-				return reject(e);
-			});
-	});
+export async function scrapeLyrics(url: string): Promise<string> {
+  const response = await fetch(url);
+
+  if (!response) {
+    return 'no lyrics found';
+  }
+
+  const text = await response.text();
+
+  if (!text) {
+    return 'text not found';
+  }
+
+  const $ = cheerio.load(text);
+  return $('.lyrics').text().trim();
 }
