@@ -1,342 +1,318 @@
-import { Message } from "discord.js";
-import { GuildPrtl, MusicData } from "../types/classes/GuildPrtl.class";
-import { CommandOptions } from "../types/classes/TypesPrtl.interface";
-import { is_url_only_channel, included_in_ignore_list } from "./guild.library";
-import { logger, isUserIgnored, isMessageDeleted, markMessageAsDeleted, messageReply } from "./help.library";
-import { remove_url, remove_ignore, set_music_data } from "./mongo.library";
-import { add_points_message } from "./user.library";
-import command_config_json from '../config.command.json';
+import { Message } from 'discord.js';
+import commandConfig from '../config.command.json';
+import { MusicData, PGuild } from '../types/classes/PGuild.class';
+import { AuthCommands, CommandOptions, NoAuthCommands, ScopeLimit } from '../types/classes/PTypes.interface';
+import { includedInPIgnores, isUrlOnlyChannel } from './guild.library';
+import { isMessageDeleted, isUserIgnored, logger, markMessageAsDeleted, messageReply } from './help.library';
+import { removeIgnore, removeURL, setMusicData } from './mongo.library';
+import { addPointsMessage } from './user.library';
 
 /*
-* Returns: true/false if processing must continue
-*/
-export async function portalPreprocessor(
-    message: Message, guild_object: GuildPrtl
-): Promise<boolean> {
-    if (!message.member) {
-        logger.error(new Error('could not get member'));
-        return true;
-    }
+ * Returns: true/false if processing must continue
+ */
+export async function portalPreprocessor(message: Message, pGuild: PGuild): Promise<boolean> {
+  if (!message.member) {
+    logger.error(new Error('could not get member'));
+    return true;
+  }
 
-    if (isUserIgnored(message.member)) {
-        if (!handleUrlChannels(message, guild_object)) {
-            if (guild_object.music_data.channel_id === message.channel.id) {
-                message.member
-                    .send('you can\'t play music when ignored')
-                    .catch((e: any) => {
-                        logger.error(new Error(`failed to send message: ${e}`));
-                    });
-
-                if (isMessageDeleted(message)) {
-                    const deletedMessage = await message
-                        .delete()
-                        .catch((e: any) => {
-                            logger.error(new Error(`failed to delete message: ${e}`));
-                        });
-
-                    if (deletedMessage) {
-                        markMessageAsDeleted(deletedMessage);
-                    }
-                }
-            }
-        }
-
-        return true;
-    } else {
-        if (await handleUrlChannels(message, guild_object)) {
-            return true;
-        }
-        else if (handleIgnoredChannels(message, guild_object)) {
-            handleRankingSystem(message, guild_object);
-            return true;
-        }
-        else if (handleMusicChannels(message, guild_object)) {
-            handleRankingSystem(message, guild_object);
-            return true;
-        } else {
-            handleRankingSystem(message, guild_object);
-
-            // if (guild_object.profanity_level !== ProfanityLevelEnum.none) {
-            //     // profanity check
-            //     const profanities = isProfane(message.content, guild_object.profanity_level);
-            //     if (profanities.length > 0) {
-            //         message
-            //             .react('🚩')
-            //             .catch((e: any) => {
-            //                 logger.error(new Error(`failed to react message: ${e}`));
-            //             });
-
-            //         message.author
-            //             .send(`try not to use profanities (${profanities.join(',')})`)
-            //             .catch(e => {
-            //                 logger.error(new Error(e));
-            //             });
-            //     }
-            // }
-
-            return false;
-        }
-    }
-}
-
-export function commandDecypher(
-    message: Message, guild_object: GuildPrtl
-): {
-    args: string[],
-    cmd: string,
-    path_to_command: string,
-    command_options: CommandOptions | undefined,
-    type: string
-} {
-    // separate command name and arguments
-    const args = message.content.slice(guild_object.prefix.length).trim().split(/ +/g);
-
-    const cmd_only = args.shift();
-    if (!cmd_only) {
-        return {
-            args: [],
-            cmd: '',
-            path_to_command: '',
-            command_options: undefined,
-            type: ''
-        };
-    }
-    const cmd = cmd_only.toLowerCase();
-
-    let path_to_command = '';
-    let command_options: CommandOptions | undefined = undefined;
-    let type = 'none';
-
-    command_config_json.some(category => {
-        command_options = category.commands.
-            find(command => command.name === cmd);
-
-        if (command_options) {
-            type = command_options.range;
-            path_to_command = category.path;
-
-            return true;
-        }
-
-        return false;
-    });
-
-    return {
-        args: args,
-        cmd: cmd,
-        path_to_command: path_to_command,
-        command_options: command_options,
-        type: type
-    };
-}
-
-export function handleRankingSystem(
-    message: Message, guild_object: GuildPrtl
-): void {
-    add_points_message(message, guild_object.member_list[0], guild_object.rank_speed)
-        .then(level => {
-            if (level) {
-                messageReply(true, message, `you reached level ${level}!`)
-                    .catch((e: any) => {
-                        logger.error(new Error(`failed to send message: ${e}`));
-                    });
-            }
-        })
-        .catch(e => {
-            logger.error(new Error(e));
+  if (isUserIgnored(message.member)) {
+    if (!handleUrlChannels(message, pGuild)) {
+      if (pGuild.musicData.channelId === message.channel.id) {
+        message.member.send('you can\'t play music when ignored').catch((e) => {
+          logger.error(new Error(`failed to send message: ${e}`));
         });
+
+        if (isMessageDeleted(message)) {
+          const deletedMessage = await message.delete().catch((e) => {
+            logger.error(new Error(`failed to delete message: ${e}`));
+          });
+
+          if (deletedMessage) {
+            markMessageAsDeleted(deletedMessage);
+          }
+        }
+      }
+    }
+
+    return true;
+  } else {
+    if (await handleUrlChannels(message, pGuild)) {
+      return true;
+    } else if (handleIgnoredChannels(message, pGuild)) {
+      handleRankingSystem(message, pGuild);
+      return true;
+    } else if (handleMusicChannels(message, pGuild)) {
+      handleRankingSystem(message, pGuild);
+      return true;
+    } else {
+      handleRankingSystem(message, pGuild);
+
+      // if (pGuild.profanityLevel !== ProfanityLevelEnum.none) {
+      //     // profanity check
+      //     const profanities = isProfane(message.content, pGuild.profanityLevel);
+      //     if (profanities.length > 0) {
+      //         message
+      //             .react('🚩')
+      //             .catch((e) => {
+      //                 logger.error(new Error(`failed to react message: ${e}`));
+      //             });
+
+      //         message.author
+      //             .send(`try not to use profanities (${profanities.join(',')})`)
+      //             .catch(e => {
+      //                 logger.error(new Error(e));
+      //             });
+      //     }
+      // }
+
+      return false;
+    }
+  }
 }
 
-export async function handleUrlChannels(
-    message: Message, guild_object: GuildPrtl
-): Promise<boolean> {
-    if (is_url_only_channel(message.channel.id, guild_object)) {
-        if (message.content === './url') {
-            remove_url(guild_object.id, message.channel.id)
-                .then(r => {
-                    messageReply(true, message, `removed url channel ${r ? 'successfully' : 'unsuccessfully'}`)
-                        .catch((e: any) => {
-                            logger.error(new Error(`failed to send message: ${e}`));
-                        });
-                })
-                .catch(e => {
-                    logger.error(new Error(`failed to remove url channel: ${e}`));
-                });
+export function commandDecipher(
+  messageContent: Message['content'],
+  pGuild: PGuild
+): {
+  cmd: AuthCommands | NoAuthCommands | undefined;
+  args: string[];
+} {
+  // separate command name and arguments
+  const args = messageContent.slice(pGuild.prefix.length).trim().split(/ +/g);
+  const commandOnly = args.shift();
+
+  if (!commandOnly) {
+    return {
+      cmd: undefined,
+      args: [],
+    };
+  }
+
+  return {
+    cmd: commandOnly.toLowerCase() as AuthCommands | NoAuthCommands,
+    args: args,
+  };
+}
+
+export function commandFetcher(
+  cmd: AuthCommands | NoAuthCommands,
+  args: string[]
+): {
+  args: string[];
+  cmd: AuthCommands | NoAuthCommands | undefined;
+  pathToCommand: string;
+  commandOptions: CommandOptions | undefined;
+  scopeLimit: ScopeLimit;
+} {
+  let pathToCommand = '';
+  let commandOptions: CommandOptions | undefined = undefined;
+  let scopeLimit = ScopeLimit.NONE;
+
+  commandConfig.some((category) => {
+    commandOptions = category.commands.find((command) => command.name === cmd) as CommandOptions;
+
+    if (commandOptions) {
+      scopeLimit = commandOptions.scopeLimit;
+      pathToCommand = category.path;
+
+      return true;
+    }
+
+    return false;
+  });
+
+  return {
+    args,
+    cmd,
+    pathToCommand,
+    commandOptions,
+    scopeLimit,
+  };
+}
+
+export function handleRankingSystem(message: Message, pGuild: PGuild): void {
+  addPointsMessage(message, pGuild.pMembers[0], pGuild.rankSpeed)
+    .then((level) => {
+      if (level) {
+        messageReply(true, message, `you reached level ${level}!`).catch((e) => {
+          logger.error(new Error(`failed to send message: ${e}`));
+        });
+      }
+    })
+    .catch((e) => {
+      logger.error(new Error(e));
+    });
+}
+
+export async function handleUrlChannels(message: Message, pGuild: PGuild): Promise<boolean> {
+  if (isUrlOnlyChannel(message.channel.id, pGuild)) {
+    if (message.content === './url') {
+      removeURL(pGuild.id, message.channel.id)
+        .then((r) => {
+          messageReply(true, message, `removed url channel ${r ? 'successfully' : 'unsuccessfully'}`).catch((e) => {
+            logger.error(new Error(`failed to send message: ${e}`));
+          });
+        })
+        .catch((e) => {
+          logger.error(new Error(`failed to remove url channel: ${e}`));
+        });
+    } else {
+      message.author.send(`${message.channel} is a url-only channel`).catch((e) => {
+        logger.error(new Error(`failed to remove url channel: ${e}`));
+      });
+
+      if (isMessageDeleted(message)) {
+        const deletedMessage = await message.delete().catch((e) => {
+          logger.error(new Error(`failed to delete message: ${e}`));
+        });
+
+        if (deletedMessage) {
+          markMessageAsDeleted(deletedMessage);
         }
-        else {
-            message.author
-                .send(`${message.channel} is a url-only channel`)
-                .catch(e => {
-                    logger.error(new Error(`failed to remove url channel: ${e}`));
-                });
+      }
+    }
 
-            if (isMessageDeleted(message)) {
-                const deletedMessage = await message
-                    .delete()
-                    .catch((e: any) => {
-                        logger.error(new Error(`failed to delete message: ${e}`));
-                    });
+    return true;
+  }
 
-                if (deletedMessage) {
-                    markMessageAsDeleted(deletedMessage);
-                }
+  return false;
+}
+
+export function handleIgnoredChannels(message: Message, pGuild: PGuild): boolean {
+  if (includedInPIgnores(message.channel.id, pGuild)) {
+    if (message.content === './ignore') {
+      removeIgnore(pGuild.id, message.channel.id)
+        .then((r) => {
+          messageReply(true, message, `removed from ignored channels ${r ? 'successfully' : 'unsuccessfully'}`).catch(
+            (e) => {
+              logger.error(new Error(`failed to send message: ${e}`));
             }
-        }
-
-        return true;
+          );
+        })
+        .catch((e) => {
+          logger.error(new Error(`failed to remove ignored channel: ${e}`));
+        });
     }
 
-    return false;
+    return true;
+  }
+
+  return false;
 }
 
-export function handleIgnoredChannels(
-    message: Message, guild_object: GuildPrtl
-): boolean {
-    if (included_in_ignore_list(message.channel.id, guild_object)) {
-        if (message.content === './ignore') {
-            remove_ignore(guild_object.id, message.channel.id)
-                .then(r => {
-                    messageReply(true, message, `removed from ignored channels ${r ? 'successfully' : 'unsuccessfully'}`)
-                        .catch((e: any) => {
-                            logger.error(new Error(`failed to send message: ${e}`));
-                        });
-                })
-                .catch(e => {
-                    logger.error(new Error(`failed to remove ignored channel: ${e}`));
-                });
-        }
-
+export function handleMusicChannels(message: Message, pGuild: PGuild): boolean {
+  if (pGuild.musicData.channelId === message.channel.id) {
+    if (message.content === './music') {
+      if (!message.guild) {
+        logger.error(new Error('failed to get guild from message'));
         return true;
+      }
+
+      const musicData = new MusicData('null', 'null', 'null', [], false);
+      setMusicData(pGuild.id, musicData)
+        .then((r) => {
+          messageReply(true, message, `removed from ignored channels ${r ? 'successfully' : 'unsuccessfully'}`).catch(
+            (e) => logger.error(new Error(`failed to send message: ${e}`))
+          );
+        })
+        .catch((e) => logger.error(new Error(`failed to remove music channel: ${e}`)));
+    } else {
+      // if (!message.guild || !message.member) {
+      //     if (message.deletable) {
+      //         message
+      //             .delete()
+      //             .catch((e) => logger.error(new Error(`failed to delete message: ${e}`)));
+      //     }
+      //     return false;
+      // }
+      // const portalVoiceConnection = client.voice?.connections
+      //     .find(c => c.channel.guild.id === message.guild?.id);
+      // if (portalVoiceConnection) {
+      //     if (!portalVoiceConnection.channel.members.has(message.member.id)) {
+      //         if (message.guild) {
+      //             const portalVoiceConnection = client.voice?.connections
+      //                 .find(c => c.channel.guild.id === message.guild?.id);
+      //             const animate = portalVoiceConnection?.dispatcher
+      //                 ? !portalVoiceConnection?.dispatcher.paused
+      //                 : false;
+      //             updateMusicMessage(
+      //                 message.guild,
+      //                 pGuild,
+      //                 pGuild.musicQueue.length > 0
+      //                     ? pGuild.musicQueue[0]
+      //                     : undefined,
+      //                 'you must be in the same channel as Portal',
+      //                 animate
+      //             ).catch(e => {
+      //                 logger.error(new Error(e));
+      //             });
+      //         }
+      //         if (message.deletable) {
+      //             message
+      //                 .delete()
+      //                 .catch((e) => {
+      //                     logger.error(new Error(`failed to send message: ${e}`));
+      //                 });
+      //         }
+      //         return false;
+      //     }
+      // }
+      // start(
+      //     voiceConnection, client, message.member.user, message,
+      //     message.guild, pGuild, message.content
+      // )
+      //     .then(r => {
+      //         if (message.guild) {
+      //             const portalVoiceConnection = client.voice?.connections
+      //                 .find(c => c.channel.guild.id === message.guild?.id);
+      //             const animate = portalVoiceConnection?.dispatcher
+      //                 ? !portalVoiceConnection?.dispatcher.paused
+      //                 : false;
+      //             updateMusicMessage(
+      //                 message.guild,
+      //                 pGuild,
+      //                 pGuild.musicQueue.length > 0
+      //                     ? pGuild.musicQueue[0]
+      //                     : undefined,
+      //                 r,
+      //                 animate
+      //             ).catch(e => {
+      //                 logger.error(new Error(e));
+      //             });
+      //         }
+      //         if (message.deletable) {
+      //             message
+      //                 .delete()
+      //                 .catch((e) => {
+      //                     logger.error(new Error(`failed to send message: ${e}`));
+      //                 });
+      //         }
+      //     })
+      //     .catch(e => {
+      //         if (message.guild) {
+      //             updateMusicMessage(
+      //                 message.guild,
+      //                 pGuild,
+      //                 pGuild.musicQueue.length > 0
+      //                     ? pGuild.musicQueue[0]
+      //                     : undefined,
+      //                 `error while starting playback: ${e}`
+      //             ).catch(e => {
+      //                 logger.error(new Error(e));
+      //             });
+      //         }
+      //         if (message.deletable) {
+      //             message
+      //                 .delete()
+      //                 .catch((e) => {
+      //                     logger.error(new Error(`failed to send message: ${e}`));
+      //                 });
+      //         }
+      //     });
     }
+    return true;
+  }
 
-    return false;
-}
-
-export function handleMusicChannels(
-    message: Message, guild_object: GuildPrtl
-): boolean {
-    if (guild_object.music_data.channel_id === message.channel.id) {
-        if (message.content === './music') {
-            if (!message.guild) {
-                logger.error(new Error(`failed to get guild from message`));
-                return true;
-            }
-
-            const music_data = new MusicData('null', 'null', 'null', [], false);
-            set_music_data(guild_object.id, music_data)
-                .then(r => {
-                    messageReply(true, message, `removed from ignored channels ${r ? 'successfully' : 'unsuccessfully'}`)
-                        .catch((e: any) => logger.error(new Error(`failed to send message: ${e}`)));
-                })
-                .catch(e => logger.error(new Error(`failed to remove music channel: ${e}`)));
-        } else {
-            // if (!message.guild || !message.member) {
-            //     if (message.deletable) {
-            //         message
-            //             .delete()
-            //             .catch((e: any) => logger.error(new Error(`failed to delete message: ${e}`)));
-            //     }
-
-            //     return false;
-            // }
-
-            // const portal_voice_connection = client.voice?.connections
-            //     .find(c => c.channel.guild.id === message.guild?.id);
-
-            // if (portal_voice_connection) {
-            //     if (!portal_voice_connection.channel.members.has(message.member.id)) {
-            //         if (message.guild) {
-            //             const portal_voice_connection = client.voice?.connections
-            //                 .find(c => c.channel.guild.id === message.guild?.id);
-
-            //             const animate = portal_voice_connection?.dispatcher
-            //                 ? !portal_voice_connection?.dispatcher.paused
-            //                 : false;
-
-            //             update_music_message(
-            //                 message.guild,
-            //                 guild_object,
-            //                 guild_object.music_queue.length > 0
-            //                     ? guild_object.music_queue[0]
-            //                     : undefined,
-            //                 'you must be in the same channel as Portal',
-            //                 animate
-            //             ).catch(e => {
-            //                 logger.error(new Error(e));
-            //             });
-            //         }
-
-            //         if (message.deletable) {
-            //             message
-            //                 .delete()
-            //                 .catch((e: any) => {
-            //                     logger.error(new Error(`failed to send message: ${e}`));
-            //                 });
-            //         }
-
-            //         return false;
-            //     }
-            // }
-
-            // start(
-            //     voice_connection, client, message.member.user, message,
-            //     message.guild, guild_object, message.content
-            // )
-            //     .then(r => {
-            //         if (message.guild) {
-            //             const portal_voice_connection = client.voice?.connections
-            //                 .find(c => c.channel.guild.id === message.guild?.id);
-
-            //             const animate = portal_voice_connection?.dispatcher
-            //                 ? !portal_voice_connection?.dispatcher.paused
-            //                 : false;
-
-            //             update_music_message(
-            //                 message.guild,
-            //                 guild_object,
-            //                 guild_object.music_queue.length > 0
-            //                     ? guild_object.music_queue[0]
-            //                     : undefined,
-            //                 r,
-            //                 animate
-            //             ).catch(e => {
-            //                 logger.error(new Error(e));
-            //             });
-            //         }
-
-            //         if (message.deletable) {
-            //             message
-            //                 .delete()
-            //                 .catch((e: any) => {
-            //                     logger.error(new Error(`failed to send message: ${e}`));
-            //                 });
-            //         }
-            //     })
-            //     .catch(e => {
-            //         if (message.guild) {
-            //             update_music_message(
-            //                 message.guild,
-            //                 guild_object,
-            //                 guild_object.music_queue.length > 0
-            //                     ? guild_object.music_queue[0]
-            //                     : undefined,
-            //                 `error while starting playback: ${e}`
-            //             ).catch(e => {
-            //                 logger.error(new Error(e));
-            //             });
-            //         }
-
-            //         if (message.deletable) {
-            //             message
-            //                 .delete()
-            //                 .catch((e: any) => {
-            //                     logger.error(new Error(`failed to send message: ${e}`));
-            //                 });
-            //         }
-            //     });
-        }
-        return true;
-    }
-
-    return false;
+  return false;
 }
