@@ -1,43 +1,43 @@
 const mod_library = require('../../build/libraries/mod.library');
 
 describe('mod_library', () => {
-    it('should return true if given word is "fuck"', () => {
-        const wordToTest = 'fuck';
+  it('should return true if given word is "fuck"', () => {
+    const wordToTest = 'fuck';
 
-        const isProfaneResult = mod_library.isProfane(wordToTest);
+    const isProfaneResult = mod_library.isProfane(wordToTest);
 
-        expect(isProfaneResult).toHaveLength(1);
-    });
+    expect(isProfaneResult).toHaveLength(1);
+  });
 
-    it('should return true if given word is "sample"', () => {
-        const wordToTest = 'sample';
+  it('should return true if given word is "sample"', () => {
+    const wordToTest = 'sample';
 
-        const isProfaneResult = mod_library.isProfane(wordToTest);
+    const isProfaneResult = mod_library.isProfane(wordToTest);
 
-        expect(isProfaneResult).toHaveLength(0);
-    });
+    expect(isProfaneResult).toHaveLength(0);
+  });
 
-    it('should return true if given word is "fuckshit"', () => {
-        const wordToTest = 'fuckshit';
+  it('should return true if given word is "fuckshit"', () => {
+    const wordToTest = 'fuckshit';
 
-        const isProfaneResult = mod_library.isProfane(wordToTest);
+    const isProfaneResult = mod_library.isProfane(wordToTest);
 
-        expect(isProfaneResult).toHaveLength(1);
-    });
+    expect(isProfaneResult).toHaveLength(1);
+  });
 
-    it('should return true if given word is "πούτσα"', () => {
-        const wordToTest = 'πούτσα';
+  it('should return true if given word is "πούτσα"', () => {
+    const wordToTest = 'πούτσα';
 
-        const isProfaneResult = mod_library.isProfane(wordToTest);
+    const isProfaneResult = mod_library.isProfane(wordToTest);
 
-        expect(isProfaneResult).toHaveLength(1);
-    });
+    expect(isProfaneResult).toHaveLength(1);
+  });
 
-    it('should return true if given word is "καλημέρα"', () => {
-        const wordToTest = 'καλημέρα';
+  it('should return true if given word is "καλημέρα"', () => {
+    const wordToTest = 'καλημέρα';
 
-        const isProfaneResult = mod_library.isProfane(wordToTest);
+    const isProfaneResult = mod_library.isProfane(wordToTest);
 
-        expect(isProfaneResult).toHaveLength(0);
-    });
+    expect(isProfaneResult).toHaveLength(0);
+  });
 });
