@@ -1,12 +1,12 @@
 // const { MockMessage, MockTextChannel } = require("jest-discordjs-mocks");
 // const { client, log_portal } = require('../src/portal');
 
-test("discord.js mock test", () => {
-    // const message = new MockMessage();
-    // const channel = new MockTextChannel();
-    // message.content = "test";
-    // message.channel = channel;
-    // callMyFunctionToTestHere(message);
+test('discord.js mock test', () => {
+  // const message = new MockMessage();
+  // const channel = new MockTextChannel();
+  // message.content = "test";
+  // message.channel = channel;
+  // callMyFunctionToTestHere(message);
 
-    expect(true).toBe(true);
+  expect(true).toBe(true);
 });
